% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_heatmap.R
\name{add_legend}
\alias{add_legend}
\title{FUNCTION_TITLE}
\usage{
add_legend(...)
}
\arguments{
\item{...}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
