% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{background_pathway_sets}
\alias{background_pathway_sets}
\title{BACKGROUND GENES PATHWAY INFORMATION}
\usage{
background_pathway_sets(
  processed_data,
  gene_info,
  sub_query,
  go,
  pathway_table,
  idep_data,
  sub_pathway_files
)
}
\description{
BACKGROUND GENES PATHWAY INFORMATION
}
