% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{build_pathway_query}
\alias{build_pathway_query}
\title{Create SQL query statement for reading pathway db}
\usage{
build_pathway_query(go, query_set)
}
\arguments{
\item{go}{Pathway category "KEGG", "GOBP"}

\item{query_set}{List of genes}
}
\value{
The SQL SELECT statement
}
\description{
Convert an ID qeury for a species from the ensembl
ID type to entrez type.
}
