% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_07_genome.R
\name{chr_filter}
\alias{chr_filter}
\title{Filter chromosome data}
\usage{
chr_filter(chr_data, regulation, chr_select)
}
\arguments{
\item{chr_data}{chromosome data}

\item{regulation}{Selected gene regulation}

\item{chr_select}{list of selected chromosomes}
}
\value{
Filtered chromosome data
}
\description{
Filters gene chromosome data by regulation and list of chromosomes selected.
"All" selected for either negates filtering. Returns data frame of filtered
chromosomal data.
}
