% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{chr_normalized_ggplot}
\alias{chr_normalized_ggplot}
\title{Creates a barplot or boxplot of the normalized data by chr}
\usage{
chr_normalized_ggplot(
  counts_data,
  sample_info,
  type = "",
  all_gene_info,
  plots_color_select = "Set1",
  use_boxplot = FALSE
)
}
\arguments{
\item{counts_data}{Matrix of raw counts from gene expression data}

\item{sample_info}{Matrix of experiment design information for grouping
samples}

\item{type}{String designating the type of data to be used in the title.
Commonly either "Raw" or "Transformed"}

\item{all_gene_info}{Gene info, including chr., gene type etc.}

\item{plots_color_select}{Vector of colors for plots (used for boxplot groups)}

\item{use_boxplot}{Logical indicating whether to use boxplot (TRUE) or
barplot (FALSE, default)}
}
\value{
A barplot or boxplot as a \code{ggplot} object
}
\description{
This function takes in either raw count or processed data and creates a
formatted plot as a \code{ggplot} object that shows the 75th percentile of
normalized expression by chromosome. By default shows barplot; can use boxplot
with jitter when user selects and sample groups are well-defined.
}
\seealso{
Other preprocess functions: 
\code{\link{chr_counts_ggplot}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{gene_counts_ggplot}()},
\code{\link{individual_plots}()},
\code{\link{mean_sd_plot}()},
\code{\link{pre_process}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{total_counts_ggplot}()}

Other plots: 
\code{\link{PCA_plot}()},
\code{\link{chr_counts_ggplot}()},
\code{\link{cor_plot}()},
\code{\link{draw_sample_tree}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{gene_counts_ggplot}()},
\code{\link{individual_plots}()},
\code{\link{k_means_elbow}()},
\code{\link{mean_sd_plot}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{sd_density}()},
\code{\link{t_SNE_plot}()},
\code{\link{total_counts_ggplot}()}
}
\concept{plots}
\concept{preprocess functions}
