% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_07_genome.R
\name{chromosome_plotly}
\alias{chromosome_plotly}
\title{MAIN PLOTLY FUNCTION}
\usage{
chromosome_plotly(
  limma,
  select_contrast,
  all_gene_info,
  ignore_non_coding,
  limma_p_val_viz,
  limma_fc_viz,
  label_gene_symbol,
  ma_window_size,
  ma_window_steps,
  ch_region_p_val
)
}
\description{
MAIN PLOTLY FUNCTION
}
