% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{clean_gene_set}
\alias{clean_gene_set}
\title{clean_gene_set}
\usage{
clean_gene_set(gene_set)
}
\arguments{
\item{gene_set}{Gene Set}
}
\value{
cleaned gene set
}
\description{
Clean up gene sets. Remove spaces and other control characters from gene names
}
\examples{
# ADD_EXAMPLES_HERE
}
