% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{clean_query}
\alias{clean_query}
\title{clean_query}
\usage{
clean_query(query_input)
}
\arguments{
\item{query_input}{string with list of genes}
}
\value{
string of cleaned gene list
}
\description{
adds tabs and new lines after each gene
}
\examples{
# ADD_EXAMPLES_HERE
}
