% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{cluster_heat_click_info}
\alias{cluster_heat_click_info}
\title{Interactive click text for subheatmap}
\usage{
cluster_heat_click_info(
  click,
  ht_sub,
  ht_sub_obj,
  ht_pos_sub,
  sub_groups,
  group_colors,
  cluster_meth,
  click_data
)
}
\arguments{
\item{click}{Click input from subheatmap}

\item{ht_sub}{Heatmap object of drawn subheatmap object}

\item{ht_sub_obj}{Heatmap object with mapping info}

\item{ht_pos_sub}{DataFrame object of position information from submap}

\item{sub_groups}{Vector of group labels from submap}

\item{group_colors}{Vector of colors for the group annotation}

\item{cluster_meth}{Integer indicating which clustering method to use 1 for
hierarchical and 2 for kmeans.}

\item{click_data}{Data matrix to get the data value from}
}
\value{
HTML code to produce a table with information
 about the selected cell.
}
\description{
Create a text output to tell the user the cell
information for their click.
}
\seealso{
Other clustering functions: 
\code{\link{create_marker_legend}()},
\code{\link{draw_sample_tree}()},
\code{\link{heat_sub}()},
\code{\link{k_means_elbow}()},
\code{\link{process_heatmap_data}()},
\code{\link{sd_density}()},
\code{\link{sub_heat_ann}()}

Other heatmaps: 
\code{\link{create_marker_legend}()},
\code{\link{heat_sub}()},
\code{\link{process_heatmap_data}()},
\code{\link{sub_heat_ann}()}
}
\concept{clustering functions}
\concept{heatmaps}
