% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kegg_pathview.R
\name{colorpanel2}
\alias{colorpanel2}
\title{Generate a color panel gradient}
\usage{
colorpanel2(n, low, mid, high)
}
\arguments{
\item{n}{Number of HEX color values to return}

\item{low}{String designating the low color}

\item{mid}{String designating the mid color}

\item{high}{String designating the high color}
}
\value{
A vector of length \code{n} with HEX color values.
}
\description{
Adapted from pathview source:
https://rdrr.io/bioc/pathview/src/R/colorpanel2.R
}
