% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{connect_convert_db_org}
\alias{connect_convert_db_org}
\title{Connect to the convertIDs database for the species and return the
objects.}
\usage{
connect_convert_db_org(datapath = DATAPATH, select_org, idep_data)
}
\arguments{
\item{datapath}{Folder path to the data file}

\item{select_org}{The slected species}

\item{idep_data}{Data object that includes org_info}
}
\value{
Database connection.
}
\description{
Create a database connection with the DBI package.
}
