% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{conversion_counts_message}
\alias{conversion_counts_message}
\title{CONVERTED COUNTS BIAS}
\usage{
conversion_counts_message(data_size, all_gene_names, n_matched)
}
\description{
CONVERTED COUNTS BIAS
}
