% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{convert_ensembl_to_entrez}
\alias{convert_ensembl_to_entrez}
\title{Convert IDs from ensembl to entrez}
\usage{
convert_ensembl_to_entrez(query, species, org_info, idep_data)
}
\arguments{
\item{query}{Vector of IDs to convert}

\item{species}{String designating the organism being analyzed}

\item{org_info}{org_info file from the list returned from
\code{link{get_idep_data}()}}

\item{idep_data}{Data object with species info for connecting to database}
}
\value{
The queried genes with converted IDs.
}
\description{
Convert an ID query for a species from the ensembl
ID type to entrez type.
}
