% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{convert_id}
\alias{convert_id}
\title{FUNCTION_TITLE}
\usage{
convert_id(
  query,
  species_choice,
  annotated_species_counts,
  select_org = "BestMatch"
)
}
\arguments{
\item{query}{DESCRIPTION.}

\item{species_choice}{DESCRIPTION.}

\item{annotated_species_counts}{DESCRIPTION.}

\item{select_org}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
