% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{cor_plot}
\alias{cor_plot}
\title{Create a correlation plot from heatmap data}
\usage{
cor_plot(data, label_pcc, heat_cols, text_col)
}
\arguments{
\item{data}{Heatmap data from \code{\link{process_heatmap_data}()}}

\item{label_pcc}{TRUE/FALSE  t0 Label with correlation coefficient}

\item{heat_cols}{Vector of colors to use in the correlation matrix}

\item{text_col}{Color to make the text labels in the plot}
}
\value{
\code{ggplot2} object as heatmap of correlation matrix
}
\description{
Creates a correlation matrix heatmap from the
heatmap data to demonstrate the correlation
between samples.
}
\seealso{
Other plots: 
\code{\link{draw_sample_tree}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{individual_plots}()},
\code{\link{k_means_elbow}()},
\code{\link{mean_sd_plot}()},
\code{\link{sd_density}()},
\code{\link{total_counts_ggplot}()}
}
\concept{plots}
