% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R, R/utils_analysis_random.R
\name{data_frame_with_list}
\alias{data_frame_with_list}
\title{GET RID OF LISTS IN A DATA FRAME}
\usage{
data_frame_with_list(data_object)

data_frame_with_list(data_object)
}
\arguments{
\item{data_object}{}
}
\value{
Data frame
}
\description{
Remove any list elements in a data frame
}
