% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{deg_click_info}
\alias{deg_click_info}
\title{DEG SUB CLICK INFO}
\usage{
deg_click_info(
  click,
  ht_sub,
  ht_sub_obj,
  ht_pos_sub,
  sub_groups,
  group_colors,
  bar,
  data
)
}
\description{
DEG SUB CLICK INFO
}
