% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{deg_deseq2}
\alias{deg_deseq2}
\title{Differential expression using DESeq2 package}
\usage{
deg_deseq2(
  raw_counts,
  max_p_limma = 0.05,
  min_fc_limma = 2,
  selected_comparisons = NULL,
  sample_info = NULL,
  model_factors = NULL,
  block_factor = NULL,
  reference_levels = NULL,
  threshold_wald_test = FALSE,
  independent_filtering = TRUE
)
}
\arguments{
\item{raw_counts}{The matrix of counts before processing for
gene expression data}

\item{max_p_limma}{Significant p-value to use for the fold-change
values}

\item{min_fc_limma}{Minimum fold-change to include in the results}

\item{selected_comparisons}{Comparisons being analyzed in the DEG
analysis}

\item{sample_info}{Experiment file information for grouping}

\item{model_factors}{Vector of selected factors and interaction terms
from the model design}

\item{block_factor}{The selected factors for batch effect}

\item{reference_levels}{Vector of reference levels to use for the
selected factors}

\item{threshold_wald_test}{whether to use threshold-based Wald test
to test null hypothesis that the absolute value of fold-change is
bigger than a value.}

\item{independent_filtering}{If yes(default), conduct independent filtering}
}
\value{
The return value is the results of the DEG analysis. These
 results are filtered and formatted by the limma_value function.
  results, a data frame with up or down regulated genes for all comparisons
  comparisons, a vectors holding comparison_names,
  exp_type, a character holding experimental design or error messages.
  top_genes, a list, each elements hold the lfc & FDR for a comparison
}
\description{
Used in the limma_value function to perform DEG analysis using the
DESeq2 package. It is not recommended to use this function on its own.
}
