% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{deg_heat_data}
\alias{deg_heat_data}
\title{HEATMAP DATA FOR DEG}
\usage{
deg_heat_data(
  limma,
  select_contrast,
  processed_data,
  contrast_samples,
  all_gene_names
)
}
\description{
HEATMAP DATA FOR DEG
}
