% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{deg_information}
\alias{deg_information}
\title{returns results from DEG analysis}
\usage{
deg_information(
  limma_value,
  gene_names,
  processed_data,
  no_id_conversion = FALSE
)
}
\arguments{
\item{limma_value}{Results from DESeq2 or limma}

\item{gene_names}{Gene IDs}

\item{processed_data}{Processed data}

\item{no_id_conversion}{If true, gene ids will not be mapped to Ensembl}
}
\value{
A list. deg_data and limma$results
}
\description{
returns results from DEG analysis
}
