% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{detect_groups}
\alias{detect_groups}
\title{detect_groups}
\usage{
detect_groups(sample_names, sample_info = NULL)
}
\arguments{
\item{sample_names}{column headings from data file or design file}

\item{sample_info}{Experiment design matrix}
}
\value{
char vector
}
\description{
Detects groups from column names in sample_info file.
}
\examples{
# ADD_EXAMPLES_HERE
}
