% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{draw_sample_tree}
\alias{draw_sample_tree}
\title{Draw a dendogram of data samples}
\usage{
draw_sample_tree(
  tree_data,
  gene_centering,
  gene_normalize,
  sample_centering,
  sample_normalize,
  hclust_funs,
  hclust_function,
  dist_funs,
  dist_function
)
}
\arguments{
\item{tree_data}{Data that has been through pre-processing}

\item{gene_centering}{TRUE/FALSE subtract mean from gene rows}

\item{gene_normalize}{TRUE/FALSE divide by SD in gene rows}

\item{sample_centering}{TRUE/FALSE subtract mean from sample columns}

\item{sample_normalize}{TRUE/FALSE divide by SD in sample columns}

\item{hclust_funs}{Clustering functions defined in idep, \code{\link{hcluster_functions}()}}

\item{hclust_function}{String of chosen clustering method}

\item{dist_funs}{Distance functions defined in idep \code{\link{dist_functions}()}}

\item{dist_function}{Selected distance function}
}
\value{
Dendogram plot of dataset samples
}
\description{
Create a clustered tree of the samples in the dataset based on hierarchical
  clustering
}
\seealso{
Other plots: 
\code{\link{PCA_plot}()},
\code{\link{chr_counts_ggplot}()},
\code{\link{chr_normalized_ggplot}()},
\code{\link{cor_plot}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{gene_counts_ggplot}()},
\code{\link{individual_plots}()},
\code{\link{k_means_elbow}()},
\code{\link{mean_sd_plot}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{sd_density}()},
\code{\link{t_SNE_plot}()},
\code{\link{total_counts_ggplot}()}

Other clustering functions: 
\code{\link{cluster_heat_click_info}()},
\code{\link{heat_sub}()},
\code{\link{heatmap_main}()},
\code{\link{k_means_elbow}()},
\code{\link{limit_annotation_factors}()},
\code{\link{process_heatmap_data}()},
\code{\link{sd_density}()}
}
\concept{clustering functions}
\concept{plots}
