% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{dynamic_range}
\alias{dynamic_range}
\title{Dynamic Range}
\usage{
dynamic_range(num_set)
}
\arguments{
\item{num_set}{Vector of values}
}
\value{
distance between 2nd largest and 2nd smallest value
}
\description{
Find the difference between 2nd largest and 2nd smallest number in set
}
\examples{
# ADD_EXAMPLES_HERE
}
