% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{eda_scatter}
\alias{eda_scatter}
\title{Scatterplot for EDA on processed data}
\usage{
eda_scatter(processed_data, plot_xaxis, plot_yaxis)
}
\arguments{
\item{processed_data}{Matrix of data that has gone through
\code{\link{prep_process}()}}

\item{plot_xaxis}{Character string indicating sample to plot on the x-axis}

\item{plot_yaxis}{Character string indicating sample to plot on the y axis}
}
\value{
A scatterplot a \code{ggplot} object
}
\description{
This function takes the data after it has been pre-processed and
creates a scatterplot of the counts for two samples that are
indicated by the user.
}
\seealso{
Other plots: 
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{individual_plots}()},
\code{\link{mean_sd_plot}()},
\code{\link{total_counts_ggplot}()}

Other preprocess functions: 
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{individual_plots}()},
\code{\link{mean_sd_plot}()},
\code{\link{pre_process}()},
\code{\link{total_counts_ggplot}()}
}
\concept{plots}
\concept{preprocess functions}
