% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_11_enrichment.R
\name{enrich_barplot}
\alias{enrich_barplot}
\title{Generate barplot for enrichment results}
\usage{
enrich_barplot(
  enrichment_dataframe,
  pathway_order,
  order_x,
  plot_size,
  plot_color,
  plot_font_size,
  plot_marker_size,
  plot_high_color,
  plot_low_color,
  chart_type,
  aspect_ratio,
  select_cluster
)
}
\arguments{
\item{enrichment_dataframe}{A data frame of pathways, P values ect.}

\item{pathway_order}{Sort pathway list}

\item{order_x}{x-axis order}

\item{plot_size}{size mapping vairalbe}

\item{plot_color}{color mapping vairable}

\item{plot_font_size}{font size}

\item{plot_marker_size}{marker size}

\item{plot_high_color}{High color}

\item{plot_low_color}{low color}

\item{chart_type}{barplot, lollipop, or dotplot}

\item{aspect_ratio}{aspect ratio of plot}

\item{select_cluster}{which cluster is selected}

\item{threshold_wald_test}{whether to use threshold-based Wald test}
}
\value{
A ggplot2 object
}
\description{
Used to translate enrichment analysis results into bar plot like those in ShinyGO
}
