% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_11_enrichment.R
\name{enrichment_network}
\alias{enrichment_network}
\title{ENRICHMENT NETWORK FUNCTION}
\usage{
enrichment_network(go_table, layout_button = 0, edge_cutoff = 5)
}
\description{
ENRICHMENT NETWORK FUNCTION
}
