% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_11_enrichment.R
\name{enrichment_tree_plot}
\alias{enrichment_tree_plot}
\title{Dendogram of enriched pathways}
\usage{
enrichment_tree_plot(
  go_table,
  group,
  right_margin = 10,
  leaf_color_choices = NULL
)
}
\arguments{
\item{go_table}{Enrichment table from the pathway analysis, the last column Genes
contains lists}

\item{group}{Selected group int the Direction column}

\item{right_margin}{Control the size of the dendogram labels}

\item{leaf_color_choices}{Vector of 2 colors for down and up regulation}
}
\value{
A dendogram plot that shows the users what pathways are
 that are enriched share genes.
}
\description{
Create a dendogram plot of the enriched pathways to illustrate
which paths contain similar genes.
}
