% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{experiment_design_txt}
\alias{experiment_design_txt}
\title{EXPERIMENT DESIGN TEXT}
\usage{
experiment_design_txt(
  sample_info,
  select_factors_model,
  select_block_factors_model,
  select_interactions
)
}
\description{
EXPERIMENT DESIGN TEXT
}
