% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{extract_fcfdr}
\alias{extract_fcfdr}
\title{Extract fold change and FDR for each comparison from limma}
\usage{
extract_fcfdr(comp, fit_contrast)
}
\arguments{
\item{comp}{Comparison}

\item{fit_contrast}{fitted contrust from limma
returned list}
}
\value{
a data frame of LFC and FDR columns
}
\description{
Extract fold change and FDR for each comparison from limma
}
