% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{extract_under}
\alias{extract_under}
\title{Replace underscore
Find any instances of "_" and replace them with a space}
\usage{
extract_under(x)
}
\arguments{
\item{x}{String}
}
\description{
Replace underscore
Find any instances of "_" and replace them with a space
}
