% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_k_means.R
\name{fct_04_k_means.R}
\alias{fct_04_k_means.R}
\title{fct_04_k_means.R This file holds all of the main data analysis functions
associated with fourth tab of the iDEP website.}
\description{
fct_04_k_means.R This file holds all of the main data analysis functions
associated with fourth tab of the iDEP website.
}
\section{fct_04_k_means.R functions}{

\code{add_legend}
}

