% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_07_deg2.R
\name{fct_07_deg2.R}
\alias{fct_07_deg2.R}
\title{fct_07_deg2.R This file holds all of the main data analysis functions
associated with seventh tab of the iDEP website.}
\description{
fct_07_deg2.R This file holds all of the main data analysis functions
associated with seventh tab of the iDEP website.
}
\section{fct_07_deg2.R functions}{

\code{short_species_names} makes short specie names for string api
}

