% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_random.R
\name{fct_analysis_random.R}
\alias{fct_analysis_random.R}
\title{fct_analysis_random.R}
\description{
fct_analysis_random.R
}
\section{fct_analysis_random.R functions}{

\code{gene_group_heatmap} heatmap with color bar define gene groups
}

