% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{fgsea_data}
\alias{fgsea_data}
\title{FGSEA DATA}
\usage{
fgsea_data(
  select_contrast,
  my_range,
  limma,
  gene_p_val_cutoff,
  gene_sets,
  absolute_fold,
  pathway_p_val_cutoff,
  n_pathway_show
)
}
\description{
FGSEA DATA
}
