% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_random.R
\name{find_contrast_samples}
\alias{find_contrast_samples}
\title{Determine samples in selected contrast}
\usage{
find_contrast_samples(
  select_contrast,
  all_sample_names,
  sample_info = NULL,
  select_factors_model = NULL,
  select_model_comprions = NULL,
  reference_levels = NULL,
  counts_deg_method = NULL,
  data_file_format = NULL
)
}
\arguments{
\item{select_contrast}{String designating the comparison from DEG analysis to
filter for the significant genes. See the 'comparison' element from the list
returned from \code{\link{limma_value}()} for options.}

\item{all_sample_names}{List of the column names of the processed gene data}

\item{sample_info}{Matrix of experiment design information for grouping
samples}

\item{select_factors_model}{List designating the selected factors for the
model expression. Example, c("P53", "Treatment"). See
\code{\link{list_factors_ui}()} for specific options.}

\item{select_model_comprions}{String designating selected comparisons to
analyze in the DEG analysis}

\item{reference_levels}{Vector of reference levels to use for the
selected factors. Should be in the form of "p53: NULL vs. WT". See
\code{\link{list_model_comparisons_ui}()} for specific options.}

\item{counts_deg_method}{Integer indicating method of DEG analysis being
performed. This should be one of 1 for limma-trend, 2 for limma-voom, and
3 for DESeq2.}

\item{data_file_format}{Integer indicating the data format. This should be
one of 1 for read counts data, 2 for normalized expression, 3 for
fold changes with adjusted P-values, or 4 for fold changes only}
}
\value{
A numeric vector that can be used to index the processed
 data and subset to only include the columns from the selected
 contrast.
}
\description{
Find the samples that are in the group of the selected
contrast. Can be used to subset the data to only include
the samples that correspond to the chosen comparison.
}
