% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{find_id_type_by_id}
\alias{find_id_type_by_id}
\title{FUNCTION_TITLE}
\usage{
find_id_type_by_id(id, id_index)
}
\arguments{
\item{id}{DESCRIPTION.}

\item{id_index}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
