% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_random.R
\name{find_overlap_gmt}
\alias{find_overlap_gmt}
\title{FUNCTION_TITLE}
\usage{
find_overlap_gmt(
  query,
  gene_set,
  min_fdr = 0.2,
  min_size = 2,
  max_size = 10000
)
}
\arguments{
\item{query}{DESCRIPTION.}

\item{gene_set}{DESCRIPTION.}

\item{min_fdr}{DESCRIPTION.}

\item{min_size}{DESCRIPTION.}

\item{max_size}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
