% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{find_species_by_id_name}
\alias{find_species_by_id_name}
\title{FUNCTION_TITLE}
\usage{
find_species_by_id_name(species_id, org_info)
}
\arguments{
\item{species_id}{DESCRIPTION.}

\item{org_info}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
