% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{find_species_id_by_ensembl}
\alias{find_species_id_by_ensembl}
\title{Find a species id by ensembl dataset name}
\usage{
find_species_id_by_ensembl(ensembl_dataset, org_info)
}
\arguments{
\item{org_info}{iDEP data org_info file}

\item{species_id}{Species ID to search the database with}
}
\value{
Only return the species name with this function.
}
\description{
Find a species in the iDEP database with an
ID.
}
