% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{gage_data}
\alias{gage_data}
\title{GAGE PATHWAY DATA}
\usage{
gage_data(
  select_go,
  select_contrast,
  min_set_size,
  max_set_size,
  limma,
  gene_p_val_cutoff,
  gene_sets,
  absolute_fold,
  pathway_p_val_cutoff,
  n_pathway_show
)
}
\description{
GAGE PATHWAY DATA
}
