% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{gene_counts_ggplot}
\alias{gene_counts_ggplot}
\title{Creates a barplot of the count of genes by type}
\usage{
gene_counts_ggplot(
  counts_data,
  sample_info,
  type = "",
  all_gene_info,
  plots_color_select
)
}
\arguments{
\item{counts_data}{Matrix of raw counts from gene expression data}

\item{sample_info}{Matrix of experiment design information for grouping
samples}

\item{type}{String designating the type of data to be used in the title.
Commonly either "Raw" or "Transformed"}

\item{all_gene_info}{Gene info, including chr., gene type etc.}

\item{plots_color_select}{Vector of colors for plots}
}
\value{
A barplot as a \code{ggplot} object
}
\description{
This function takes in either raw count or processed data and creates a
formatted barplot as a \code{ggplot} object that shows the number
of genes mapped to each sample in millions. This function is only used for
read counts data.
}
\seealso{
Other preprocess functions: 
\code{\link{chr_counts_ggplot}()},
\code{\link{chr_normalized_ggplot}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{individual_plots}()},
\code{\link{mean_sd_plot}()},
\code{\link{pre_process}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{total_counts_ggplot}()}

Other plots: 
\code{\link{PCA_plot}()},
\code{\link{chr_counts_ggplot}()},
\code{\link{chr_normalized_ggplot}()},
\code{\link{cor_plot}()},
\code{\link{draw_sample_tree}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{individual_plots}()},
\code{\link{k_means_elbow}()},
\code{\link{mean_sd_plot}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{sd_density}()},
\code{\link{t_SNE_plot}()},
\code{\link{total_counts_ggplot}()}
}
\concept{plots}
\concept{preprocess functions}
