% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_random.R
\name{gene_group_heatmap}
\alias{gene_group_heatmap}
\title{FUNCTION_TITLE}
\usage{
gene_group_heatmap(
  x,
  bar = NULL,
  n = -1,
  mycolor = 1,
  clusterNames = NULL,
  sideColors = NULL
)
}
\arguments{
\item{x}{DESCRIPTION.}

\item{bar}{DESCRIPTION.}

\item{n}{DESCRIPTION.}

\item{mycolor}{DESCRIPTION.}

\item{clusterNames}{DESCRIPTION.}

\item{sideColors}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
