% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_01_load_data.R
\name{gene_info}
\alias{gene_info}
\title{Retrieve detailed info on genes}
\usage{
gene_info(converted, select_org, idep_data)
}
\arguments{
\item{converted}{List returned \code{\link{convert_id}()}. Contains
information about the gene IDs for the matched species.}

\item{select_org}{String indicating selected organism for the expression
data.Default is "BestMatch."}

\item{idep_data}{Data files from the database obtained with
\code{\link{get_idep_data}()}}
}
\value{
A data frame containing information on all ensembl IDs for the
 matched species. Usually a very large data frame due to the amount
 of IDs that the data base contains.
}
\description{
This function retrieves detailed gene information from the
database for the matched species.
}
\seealso{
Other load data functions: 
\code{\link{convert_data}()},
\code{\link{get_all_gene_names}()},
\code{\link{input_data}()}
}
\concept{load data functions}
