% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{generate_colors}
\alias{generate_colors}
\title{Generate colors for plots, solving the problem of not having enough colors for some selections}
\usage{
generate_colors(n, palette_name = "Set1")
}
\arguments{
\item{n}{Number of desired colors}

\item{palette_name}{Selected color set, i.e. Set1, Set2, etc.}
}
\value{
a list of n colors, even if there are fewer colors in the palette
}
\description{
Creates a message about the size of the counts
data and the amount of IDs that were converted.
}
