% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Gets plot dimensions from session$clientdata}
\usage{
get_dimensions(client_data, plot_name, tab)
}
\arguments{
\item{data}{Data that has been through pre-processing}

\item{sample_info}{Design Matrix}
}
\value{
Formatted plot generated with PCAtools package
}
\description{
Gets plot dimensions from session$clientdata
}
