% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_random.R
\name{get_gene_info}
\alias{get_gene_info}
\title{Retrieve detailed gene information}
\usage{
get_gene_info(converted, select_org, idep_data)
}
\arguments{
\item{converted}{List of converted gene information  from
\code{\link{convert_id}()}}

\item{select_org}{String designating the organism being analyzed.}

\item{idep_data}{Data object returned by \code{\link{get_idep_data}()}
containing database configuration.}
}
\value{
A data frame
}
\description{
Retrieve detailed gene information
}
