% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_heatmap.R
\name{get_heatmap_colors}
\alias{get_heatmap_colors}
\title{Get color choice list for the heatmap}
\usage{
get_heatmap_colors()
}
\value{
Returns a list containing color choices and a data frame
of hex color values.
}
\description{
get_heatmap_colors
}
\details{
Calling the function will provide a list of heatmap colors
for the select input bar to use and gives the colors that the
heatmap will use for the selected color set.
}
