% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{get_idep_data}
\alias{get_idep_data}
\title{Load iDEP data}
\usage{
get_idep_data(datapath = DATAPATH)
}
\arguments{
\item{datapath}{Folder path to the iDEP data}
}
\value{
A list containing:
\describe{
  \item{demo_file_info}{Information on bundled demo datasets, including
  paths to expression and design files.}
  \item{org_info}{Species metadata sourced from the Ensembl
  orgInfo table.}
}
}
\description{
Use this function call to load data that is
used in other functions.
}
