% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_09_network.R
\name{get_network}
\alias{get_network}
\title{Network of top genes}
\usage{
get_network(
  select_wgcna_module,
  wgcna,
  top_genes_network,
  select_org,
  all_gene_info,
  edge_threshold
)
}
\arguments{
\item{select_wgcna_module}{The module to create a plot of the top genes for,
options can be found with the \code{get_wgcna_modules} function}

\item{wgcna}{List returned from the \code{get_wgcna}}

\item{top_genes_network}{Number of genes to include in the network plot}

\item{select_org}{Organism the expression data is for}

\item{all_gene_info}{Gene info that was found from querying the database}

\item{edge_threshold}{Value from 1 to 0.1 (0.4 recommended)}
}
\value{
An adjacency matrix of the top genes in the selected module
}
\description{
Create a network plot of the top genes found with the WGCNA package.
}
