% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_09_network.R
\name{get_network_plot}
\alias{get_network_plot}
\title{MODULE PLOT}
\usage{
get_network_plot(
  select_wgcna_module,
  wgcna,
  top_genes_network,
  select_go,
  select_org,
  all_gene_info,
  edge_threshold
)
}
\description{
MODULE PLOT
}
