% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{get_pathway_list_data}
\alias{get_pathway_list_data}
\title{Create pathway table with gene sets}
\usage{
get_pathway_list_data(
  pathway_method,
  gage_pathway_data,
  fgsea_pathway_data,
  pgsea_plot_data,
  pgsea_plot_all_samples_data,
  go,
  select_org,
  gene_info,
  gene_sets
)
}
\arguments{
\item{pathway_method}{Integer indicating which pathway method to use. Should
be one of 1 for "GAGE", 2 = "PGSEA", 3 for "GSEA (preranded fgsea)", 4
for "PGSEA w/ all samples", and 5 for "ReactomePA".}

\item{gage_pathway_data}{Matrix returned from \code{\link{gage_data}()}}

\item{fgsea_pathway_data}{Matrix returned from \code{\link{fgsea_data}()}}

\item{pgsea_plot_data}{Matrix returned from
\code{\link{get_pgsea_plot_data}()}}

\item{pgsea_plot_all_samples_data}{Matrix returned from
\code{\link{get_pgsea_plot_all_samples_data}()}}

\item{go}{String designating the section of the database to query for
pathway analysis. See \code{\link{gmt_category}()} for choices.}

\item{select_org}{String designating which organism is being analyzed}

\item{gene_info}{Matrix returned from \code{\link{gene_info}()} function, all
gene info from the database query with the User gene IDs}

\item{gene_sets}{List of vectors with each vector being the
set of genes that correspond to a particular pathway in
the database \code{\link{read_gene_sets function}()}}
}
\value{
A data frame with the pathway analysis statistics and
 the gene sets for each significantly enriched pathway.
}
\description{
Create a data frame of significant pathways and their analysis
values. Also add a column that contains the gene sets for the
pathway.
}
\seealso{
Other pathway functions: 
\code{\link{fgsea_data}()},
\code{\link{find_overlap}()},
\code{\link{gage_data}()},
\code{\link{get_pgsea_plot_all_samples_data}()},
\code{\link{get_pgsea_plot_data}()},
\code{\link{kegg_pathway}()},
\code{\link{pathway_select_data}()},
\code{\link{pgsea_data}()},
\code{\link{pgsea_plot_all}()},
\code{\link{plot_pgsea}()},
\code{\link{reactome_data}()},
\code{\link{remove_pathway_id}()}
}
\concept{pathway functions}
