% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{get_pc}
\alias{get_pc}
\title{Prepare PC data}
\usage{
get_pc(data, sample_info)
}
\arguments{
\item{data}{Data that has been through pre-processing}

\item{sample_info}{Matrix array with experiment info}
}
\value{
pca data ready for plotting
}
\description{
Prepare PC data
}
