% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{get_pc_variance}
\alias{get_pc_variance}
\title{Get variance percentages}
\usage{
get_pc_variance(data)
}
\arguments{
\item{data}{Data}
}
\value{
importance of each pc
}
\description{
Get variance percentages
}
