% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{get_plot_height}
\alias{get_plot_height}
\title{Gets plot height size from session$clientdata}
\usage{
get_plot_height(client_data, plot_name, tab)
}
\arguments{
\item{client_data}{session info from shiny}

\item{plot_name}{Name of plot object}

\item{tab}{Tab name}
}
\value{
height in inches of plot
}
\description{
Gets plot height size from session$clientdata
}
\seealso{
Other PCA functions: 
\code{\link{MDS_plot}()},
\code{\link{PCA_Scree}()},
\code{\link{PCA_biplot}()},
\code{\link{PCA_plot_3d}()},
\code{\link{PCA_plot}()},
\code{\link{PCAtools_eigencorplot}()},
\code{\link{get_plot_width}()},
\code{\link{pc_factor_correlation}()},
\code{\link{t_SNE_plot}()}
}
\concept{PCA functions}
