% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{get_tukey_data}
\alias{get_tukey_data}
\title{Get TukeyHSD Result Data}
\usage{
get_tukey_data(gene_data, sample_info, selected_gene = NULL, summarized)
}
\arguments{
\item{gene_data}{data frame of gene expression data}

\item{sample_info}{additional sample info from experimental design file}

\item{selected_gene}{list of genes selected for analysis; Only required if 
summarized = FALSE}

\item{summarized}{TRUE/FALSE whether data has been summarized}
}
\value{
Data frame containing TukeyHSD results
}
\description{
Takes a data frame of gene expression data and runs TukeyHSD tests between
sample groups within each gene. Returns as a data frame.
}
