% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{get_x_axis_label_size}
\alias{get_x_axis_label_size}
\title{Get x-axis label size based on number of samples}
\usage{
get_x_axis_label_size(n_samples)
}
\arguments{
\item{n_samples}{Number of samples in the data}
}
\value{
Numeric value for x-axis label size
}
\description{
Get x-axis label size based on number of samples
}
\seealso{
Other preprocess functions: 
\code{\link{chr_counts_ggplot}()},
\code{\link{chr_normalized_ggplot}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{gene_counts_ggplot}()},
\code{\link{individual_plots}()},
\code{\link{mean_sd_plot}()},
\code{\link{pre_process}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{total_counts_ggplot}()}
}
\concept{preprocess functions}
