% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{\code{ggplot} colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number of colors to return}
}
\value{
Vector of hex color codes for a plot.
}
\description{
Create a vector of hex colors to be used in \code{ggplot}s
}
