% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{gmt_category}
\alias{gmt_category}
\title{Database choices for the converted IDs}
\usage{
gmt_category(converted, converted_data, select_org, gmt_file, idep_data)
}
\description{
Database choices for the converted IDs
}
