% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{hcluster_functions}
\alias{hcluster_functions}
\title{List of method functions for hierarchical clustering}
\usage{
hcluster_functions()
}
\value{
A list of functions
}
\description{
Returns a list of functions for hierarchical clustering
including average linkage, ward.d, ward.d2, single linkage, mcquitty,
median, and centroid.
}
