% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R, R/utils_analysis_random.R
\name{heat_click_info}
\alias{heat_click_info}
\title{Interactive click text for subheatmap}
\usage{
heat_click_info(
  click,
  ht_sub,
  ht_sub_obj,
  ht_pos_sub,
  sub_groups,
  group_colors,
  data
)

heat_click_info(
  click,
  ht_sub,
  ht_sub_obj,
  ht_pos_sub,
  sub_groups,
  group_colors,
  data
)
}
\arguments{
\item{click}{Information from the User clicking on a cell of
the sub-heatmap}

\item{ht_sub}{The drawn sub-heatmap}

\item{ht_sub_obj}{The sub-heatmap ComplexHeatmap object}

\item{ht_pos_sub}{Position information for the sub-heatmap}

\item{sub_groups}{Vector of the groups that the samples
belong to}

\item{group_colors}{color of the top annotation that
is used for each group}

\item{data}{Sub data matrix that is plotted in the sub-heatmap}

\item{cluster_meth}{Type of clustering being performed}

\item{click_data}{Data matrix to get the data value from}
}
\value{
HTML code to produce a table with information
 about the selected cell.

HTML code that will be used in the shiny UI to tell
 the user the information of the cell they selected.
}
\description{
Create a text output to tell the user the cell
information for their click.

Create HTML code for a cell of information on the cell of the
sub-heatmap that the User clicks on. The cell contains the
expression value, the sample, the gene, and the group.
}
