% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{heat_gene_select_ui}
\alias{heat_gene_select_ui}
\title{UI component to customize gene labels}
\usage{
heat_gene_select_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
Shiny module
}
\description{
This component contains an action button to activate the pop-up modal to
select how many genes are displayed on the heat map and what method to
choose them.
}
