% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{heat_sub}
\alias{heat_sub}
\title{Draw sub heatmap from brush input}
\usage{
heat_sub(
  ht_brush,
  ht,
  ht_pos_main,
  heatmap_data,
  sample_info,
  select_factors_heatmap,
  cluster_meth
)
}
\arguments{
\item{ht_brush}{Brush input from the main heatmap}

\item{ht}{Main heatmap object}

\item{ht_pos_main}{Position of brush on main heatmap}

\item{heatmap_data}{Matrix of data for the heatmap from 
\code{\link{process_heatmap_data}()}}

\item{sample_info}{Matrix of experiment design file information}

\item{select_factors_heatmap}{Group design to label by}

\item{cluster_meth}{Integer indicating which clustering method to use 1 for 
hierarchical and 2 for kmeans.}
}
\value{
A list containing a Heatmap from the brush selection
 of the main heatmap, the submap data matrix, the groups for
 the submap, the submap legend, and data for the click info.
}
\description{
Use the brush input from the main heatmap to
create a larger subheatmap.
}
\seealso{
Other heatmaps: 
\code{\link{cluster_heat_click_info}()},
\code{\link{heatmap_main}()},
\code{\link{process_heatmap_data}()},
\code{\link{sub_heat_ann}()}

Other clustering: 
\code{\link{cluster_heat_click_info}()},
\code{\link{draw_sample_tree}()},
\code{\link{heatmap_main}()},
\code{\link{k_means_elbow}()},
\code{\link{process_heatmap_data}()},
\code{\link{sd_density}()},
\code{\link{sub_heat_ann}()}
}
\concept{clustering}
\concept{heatmaps}
