% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{hyperText}
\alias{hyperText}
\title{Convert regular text to hypertext with links}
\usage{
hyperText(textVector, urlVector)
}
\arguments{
\item{textVector}{A vector of characters}

\item{urlVector}{a set of URLs}

\item{click}{Information from the User clicking on a cell of
the sub-heatmap}
}
\value{
HTML format hyper text
}
\description{
Convert regular text to hypertext with links
}
