% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{individual_plots}
\alias{individual_plots}
\title{INDIVDUAL GENE PLOTS}
\usage{
individual_plots(
  merged_data,
  sample_info,
  selected_gene,
  gene_plot_box,
  use_sd,
  select_org
)
}
\description{
INDIVDUAL GENE PLOTS
}
