% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{k_means_elbow}
\alias{k_means_elbow}
\title{Draw an elbow plot for k-cluster selection}
\usage{
k_means_elbow(heatmap_data)
}
\arguments{
\item{heatmap_data}{Matrix of processed heatmap data from
\code{\link{process_heatmap_data}()}}
}
\value{
\code{ggplot2} object of formatted elbow plot
}
\description{
This function takes in the processed heatmap data and
creates an elbow plot to guide the selection of the
number of clusters to create
}
\seealso{
Other plots: 
\code{\link{cor_plot}()},
\code{\link{draw_sample_tree}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{individual_plots}()},
\code{\link{mean_sd_plot}()},
\code{\link{sd_density}()},
\code{\link{total_counts_ggplot}()}

Other clustering functions: 
\code{\link{cluster_heat_click_info}()},
\code{\link{draw_sample_tree}()},
\code{\link{heat_sub}()},
\code{\link{heatmap_main}()},
\code{\link{process_heatmap_data}()},
\code{\link{sd_density}()},
\code{\link{sub_heat_ann}()}
}
\concept{clustering functions}
\concept{plots}
