% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{kegg_pathway_id}
\alias{kegg_pathway_id}
\title{Given a KEGG pathway description, find pathway ids}
\usage{
kegg_pathway_id(
  pathway_description,
  Species,
  GO,
  select_org,
  gmt_files,
  org_info
)
}
\description{
Given a KEGG pathway description, find pathway ids
}
