% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{limit_annotation_factors}
\alias{limit_annotation_factors}
\title{Create annotation for shiny subheatmap}
\usage{
limit_annotation_factors(sample_info, max_factors = 3)
}
\arguments{
\item{sample_info}{Matrix of experiment design information from load data}

\item{data}{Matrix of heatmap data}

\item{select_factors_heatmap}{Factor to group by in the samples.
"All factors" will use up to the first three columns of the sample information.}

\item{group_pal}{Named list of colors and their corresponding categories}

\item{sample_color}{Selected colorspace color palette}

\item{use_letter_overlay}{Logical flag to overlay uppercase letters on sample annotations.}
}
\value{
A list containing a ComplexHeatmap annotation object,
 a ComplexHeatmap legend, list of groups, and list of group colors.
}
\description{
Use the heatmap data to make an annotation for the
submap that will also show the legend
}
\seealso{
Other clustering functions: 
\code{\link{cluster_heat_click_info}()},
\code{\link{draw_sample_tree}()},
\code{\link{heat_sub}()},
\code{\link{heatmap_main}()},
\code{\link{k_means_elbow}()},
\code{\link{process_heatmap_data}()},
\code{\link{sd_density}()}

Other heatmaps: 
\code{\link{cluster_heat_click_info}()},
\code{\link{heat_sub}()},
\code{\link{heatmap_main}()},
\code{\link{process_heatmap_data}()}
}
\concept{clustering functions}
\concept{heatmaps}
