% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{limma_value}
\alias{limma_value}
\title{LIMMA REACTIVE VALUE}
\usage{
limma_value(
  data_file_format,
  counts_deg_method,
  raw_counts,
  limma_p_val,
  limma_fc,
  select_model_comprions,
  sample_info,
  select_factors_model,
  select_interactions,
  select_block_factors_model,
  factor_reference_levels,
  processed_data,
  counts_log_start,
  p_vals
)
}
\description{
LIMMA REACTIVE VALUE
}
