% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{list_block_factors_ui}
\alias{list_block_factors_ui}
\title{Get the block factor choices}
\usage{
list_block_factors_ui(sample_info, select_factors_model)
}
\arguments{
\item{sample_info}{Matrix of experiment design information for grouping}

\item{select_factors_model}{Vector of strings indicating the selected factors
for the model expression, see \code{\link{list_factors_ui}()} for options}
}
\value{
This function returns a vector of choices for a batch
 effect or paired samples.
}
\description{
This function uses the sample_info file and the selected
factors for the model to create a selection for the batch
effect. Returns a vector that turns into a checkbox for
the User.
}
\seealso{
Other DEG functions: 
\code{\link{change_names}()},
\code{\link{deg_heat_data}()},
\code{\link{deg_information}()},
\code{\link{experiment_design_txt}()},
\code{\link{limma_value}()},
\code{\link{list_factors_ui}()},
\code{\link{list_interaction_terms_ui}()},
\code{\link{list_model_comparisons_ui}()},
\code{\link{mod_label_server}()},
\code{\link{mod_label_ui}()},
\code{\link{plot_deg_scatter}()},
\code{\link{plot_ma}()},
\code{\link{plot_volcano}()},
\code{\link{select_reference_levels_ui}()},
\code{\link{volcano_data}()}
}
\concept{DEG functions}
