% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{list_interaction_terms_ui}
\alias{list_interaction_terms_ui}
\title{List model interaction terms}
\usage{
list_interaction_terms_ui(sample_info, select_factors_model)
}
\arguments{
\item{sample_info}{Matrix of experiment design information for grouping}

\item{select_factors_model}{Vector of strings indicating the selected factors
for the model expression, see \code{\link{list_factors_ui}()} for options}
}
\value{
Returns a character string of an interaction term
 between the selected factors. Used in a checkbox for the
 User to create a model expression
}
\description{
This functions uses the sample info file and the selected
model factors to create interaction terms to be used in the
DEG process.
}
\seealso{
Other DEG functions: 
\code{\link{change_names}()},
\code{\link{deg_heat_data}()},
\code{\link{deg_information}()},
\code{\link{experiment_design_txt}()},
\code{\link{limma_value}()},
\code{\link{list_block_factors_ui}()},
\code{\link{list_factors_ui}()},
\code{\link{list_model_comparisons_ui}()},
\code{\link{mod_label_server}()},
\code{\link{mod_label_ui}()},
\code{\link{plot_deg_scatter}()},
\code{\link{plot_ma}()},
\code{\link{plot_volcano}()},
\code{\link{select_reference_levels_ui}()},
\code{\link{volcano_data}()}
}
\concept{DEG functions}
