% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{list_model_comparisons_ui}
\alias{list_model_comparisons_ui}
\title{Create model comparisons choices}
\usage{
list_model_comparisons_ui(sample_info, select_factors_model, processed_data)
}
\arguments{
\item{sample_info}{Matrix of experiment design information for grouping}

\item{select_factors_model}{Vector of strings indicating the selected factors
for the model expression, see \code{\link{list_factors_ui}()} for options}

\item{processed_data}{Matrix of gene data that has been through the 
\code{\link{pre_process}()}}
}
\value{
Returns a list containing a vector of choices and a
 title for the UI element.
}
\description{
This function uses the sample_info file and the selected
factors to create a list of options for model comparisons.
Changes with the input of select_factors_model. If there
is no selected factor then it defaults to comparisons that
can be created from the processed data.
}
\seealso{
Other DEG functions: 
\code{\link{change_names}()},
\code{\link{deg_information}()},
\code{\link{experiment_design_txt}()},
\code{\link{limma_value}()},
\code{\link{list_block_factors_ui}()},
\code{\link{list_factors_ui}()},
\code{\link{list_interaction_terms_ui}()},
\code{\link{mod_label_server}()},
\code{\link{mod_label_ui}()},
\code{\link{plot_deg_scatter}()},
\code{\link{select_reference_levels_ui}()}
}
\concept{DEG functions}
