% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{make_pull_down_menu}
\alias{make_pull_down_menu}
\title{Create a pull down menu}
\usage{
make_pull_down_menu(funs)
}
\arguments{
\item{funs}{List of functions}
}
\value{
names of a set
}
\description{
Create a pull down menu
}
