% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_01_load_data.R
\name{max_fun}
\alias{max_fun}
\title{Calculate max of vector}
\usage{
max_fun(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a number, or NA if all the numbers are NA
}
\description{
This function refines the max function to handle missing value, used 
with the aggregate function
}
