% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{MDS_plot}
\alias{MDS_plot}
\title{MDS FUNCTION}
\usage{
MDS_plot(data, sample_info, selected_shape, selected_color)
}
\arguments{
\item{data}{Data that has been through pre-processing}

\item{sample_info}{Matrix array with experiment info}

\item{selected_shape}{shape}

\item{selected_color}{color}
}
\value{
Formatted PCA plot
}
\description{
Draw a MDS plot
}
