% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_01_load_data.R
\name{mean_fun}
\alias{mean_fun}
\title{Calculate mean of vector}
\usage{
mean_fun(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a number, or NA if all the numbers are NA
}
\description{
This function refines the mean function to handle missing value, used 
with the aggregate function
}
