% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{mean_sd_plot}
\alias{mean_sd_plot}
\title{Mean vs. Standard Deviation plot}
\usage{
mean_sd_plot(processed_data, rank, heat_cols)
}
\arguments{
\item{processed_data}{Matrix of data that has gone through
\code{\link{pre_process}()}}

\item{rank}{TRUE/FALSE whether to use the rank of the mean or not}

\item{heat_cols}{Heat color to use with black in the plot}
}
\value{
A formatted \code{ggplot} hexplot of the mean and standard
 deviation of the processed data
}
\description{
Create a plot that shows the standard deviation as the
Y-axis across the mean of the counts data on the X-axis.
Option to make the X-axis the rank of the mean which
does a better job showing the spread of the data.
}
\seealso{
Other plots: 
\code{\link{PCA_plot}()},
\code{\link{chr_counts_ggplot}()},
\code{\link{chr_normalized_ggplot}()},
\code{\link{cor_plot}()},
\code{\link{draw_sample_tree}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{gene_counts_ggplot}()},
\code{\link{individual_plots}()},
\code{\link{k_means_elbow}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{sd_density}()},
\code{\link{t_SNE_plot}()},
\code{\link{total_counts_ggplot}()}

Other preprocess functions: 
\code{\link{chr_counts_ggplot}()},
\code{\link{chr_normalized_ggplot}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{gene_counts_ggplot}()},
\code{\link{individual_plots}()},
\code{\link{pre_process}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{total_counts_ggplot}()}
}
\concept{plots}
\concept{preprocess functions}
