% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_01_load_data.R
\name{median_fun}
\alias{median_fun}
\title{Calculate median of vector}
\usage{
median_fun(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a number, or NA if all the numbers are NA
}
\description{
This function refines the median function to handle missing value, used 
with the aggregate function
}
