% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{merge_data}
\alias{merge_data}
\title{Merge data from gene info and processed data}
\usage{
merge_data(select_org, all_gene_info, processed_data)
}
\arguments{
\item{select_org}{Selected species that the expression data is from}

\item{all_gene_info}{Data containing the ensembl to gene name mapping}

\item{processed_data}{Data that has gone through the pre_processing}
}
\value{

}
\description{
This function takes in the gene info data and merges it
with the data that has gone through the processing fcn.
The returned data contains the gene names as well as the
ensembl id in the first two columns.
}
