% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{mod_label_server}
\alias{mod_label_server}
\title{Server component to customize gene labels}
\usage{
mod_label_server(id, data_list, method = c("volcano", "ma"))
}
\arguments{
\item{id}{Namespace ID}

\item{data_list}{List of gene data from \code{volcano_data()}}

\item{method}{String designating if the results are for the volcano plot or
ma plot}
}
\value{
A shiny module.
}
\description{
This component contains the pop-up modal and data processing to customize
 how genes are labeled on the volcano or ma plot. Users have the option to
 not label genes, label specific genes, label top n genes by a certain value,
 or label genes above certain threshold.
}
