% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{mod_label_ui}
\alias{mod_label_ui}
\title{UI component to customize gene labels}
\usage{
mod_label_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
Shiny module
}
\description{
This component contains an action button to activate the pop-up modal to
customize how genes are labeled on the volcano or ma plot.
}
