% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{path_click_info}
\alias{path_click_info}
\title{PATH SUB CLICK INFO}
\usage{
path_click_info(
  click,
  ht_sub,
  ht_sub_obj,
  ht_pos_sub,
  sub_groups,
  group_colors,
  data
)
}
\description{
PATH SUB CLICK INFO
}
