% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{pathway_heatmap}
\alias{pathway_heatmap}
\title{Heatmap of the expression data for chosen pathway}
\usage{
pathway_heatmap(data, heatmap_color_select)
}
\arguments{
\item{data}{Data that has been filtered with the
pathway_select_data function}

\item{heatmap_color_select}{Vector of colors to use for the fill
in the heatmap}
}
\value{
A drawn ComplexHeatmap.
}
\description{
Create a ComplexHeatmap from the sub matrix that was found in
the pathway_select_data function. The heatmap will demonstrate
the gene expression difference across groups for a chosen
pathway.
}
