% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{pathway_select_data}
\alias{pathway_select_data}
\title{Get data from genes in selected pathway}
\usage{
pathway_select_data(
  sig_pathways,
  gene_sets,
  contrast_samples,
  data,
  select_org,
  all_gene_names,
  deg,
  select_contrast = NULL
)
}
\arguments{
\item{sig_pathways}{Description of the pathway for which to
obtain the gene expression data}

\item{gene_sets}{List of vectors with each vector being the
set of genes that correspond to a particular pathway in
the database. See list returned from
\code{\link{read_gene_sets}()}}

\item{contrast_samples}{Vector of sample columns that correspond to the
selected comparison}

\item{data}{Matrix of gene data that has been through
\code{\link{pre_process}()}}

\item{select_org}{String designating the organism being analyzed}

\item{all_gene_names}{Matrix of all the matched and converted
gene IDs from \code{\link{get_all_gene_names}()}}

\item{deg}{data frame of Stats gene regulation results - i.e.-1, 0 , 1 for
every gene}

\item{select_contrast}{Selected comparison/contrast name to extract regulation
status from (optional, defaults to first column if deg has multiple columns)}
}
\value{
Sub-data matrix from the processed data. Only contains
 genes from the selected pathway and samples that correspond to
 the comparison being analyzed.
}
\description{
Return a data matrix that is a subset of the processed data and
only contains genes that are in the gene set of the desired
pathway.
}
\seealso{
Other pathway functions: 
\code{\link{fgsea_data}()},
\code{\link{find_overlap}()},
\code{\link{gage_data}()},
\code{\link{get_gsva_plot_data}()},
\code{\link{get_pathway_list_data}()},
\code{\link{get_pgsea_plot_all_samples_data}()},
\code{\link{get_pgsea_plot_data}()},
\code{\link{gsva_data}()},
\code{\link{kegg_pathway}()},
\code{\link{pgsea_data}()},
\code{\link{pgsea_plot_all}()},
\code{\link{plot_gsva}()},
\code{\link{plot_pgsea}()},
\code{\link{reactome_data}()},
\code{\link{remove_pathway_id}()},
\code{\link{remove_pathway_id_second}()}
}
\concept{pathway functions}
