% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{pathway_source_info}
\alias{pathway_source_info}
\title{retrieve source info for pathway db}
\usage{
pathway_source_info(pathway_file, go, select_org, idep_data)
}
\arguments{
\item{pathway_file}{pathway file}

\item{go}{Pathway category "KEGG", "GOBP"}

\item{select_org}{selected species}

\item{idep_data}{Data object with species info for connecting to database}
}
\value{
A vector with the source information
}
\description{
retrieve source info for pathway db
}
