% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{pc_factor_correlation}
\alias{pc_factor_correlation}
\title{Correlations Between Principle Components and Factors}
\usage{
pc_factor_correlation(data, sample_info)
}
\arguments{
\item{data}{Data that has been through pre-processing}

\item{sample_info}{Matrix array with experiment info}
}
\value{
text with correlation
}
\description{
Desc
}
\seealso{
Other PCA functions: 
\code{\link{MDS_plot}()},
\code{\link{PCA_Scree}()},
\code{\link{PCA_biplot}()},
\code{\link{PCA_plot_3d}()},
\code{\link{PCA_plot}()},
\code{\link{PCAtools_eigencorplot}()},
\code{\link{get_plot_height}()},
\code{\link{get_plot_width}()},
\code{\link{t_SNE_plot}()}
}
\concept{PCA functions}
