% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{PCA_biplot}
\alias{PCA_biplot}
\title{Principal Component Analysis with PCAtools package}
\usage{
PCA_biplot(
  data,
  sample_info,
  select_gene_id = "symbol",
  all_gene_names,
  selected_x = "PC1",
  selected_y = "PC2",
  encircle = TRUE,
  encircleFill = TRUE,
  showLoadings = TRUE,
  pointlabs = TRUE,
  point_size = 4,
  ui_color = NULL,
  ui_shape = NULL
)
}
\arguments{
\item{data}{Data that has been through pre-processing}

\item{sample_info}{Matrix array with experiment info}

\item{select_gene_id}{gene name option}

\item{all_gene_names}{gene names}

\item{selected_x}{x axis selection eg "PC1"}

\item{selected_y}{y axis selection}

\item{encircle}{draw shapes in plot}

\item{encircleFill}{fill shapes in plot}

\item{showLoadings}{draw gene vectors onto plot}

\item{pointlabs}{show column names on points}

\item{point_size}{point size}

\item{ui_color}{color}

\item{ui_shape}{shape}
}
\value{
Formatted PCA plot using PCAtools package
}
\description{
Draw a PCA plot using PCAtools package
}
\seealso{
Other PCA functions: 
\code{\link{MDS_plot}()},
\code{\link{PCA_Scree}()},
\code{\link{PCA_plot_3d}()},
\code{\link{PCA_plot}()},
\code{\link{PCAtools_eigencorplot}()},
\code{\link{get_plot_height}()},
\code{\link{get_plot_width}()},
\code{\link{pc_factor_correlation}()},
\code{\link{t_SNE_plot}()}
}
\concept{PCA functions}
