% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{PCA_plot}
\alias{PCA_plot}
\title{Principal component analysis plot}
\usage{
PCA_plot(
  data,
  sample_info,
  PCAx = 1,
  PCAy = 2,
  selected_color = "Names",
  selected_shape = "Names"
)
}
\arguments{
\item{data}{Matrix of gene data that has been through pre-processing}

\item{sample_info}{Matrix of sample information from experiment design file}

\item{PCAx}{Integer designating the PC to be plotted on the x axis}

\item{PCAy}{Integer designating the PC to be plotted on on the y axis}

\item{selected_color}{String designating factor to color points by}

\item{selected_shape}{String designating factor to add shape to points by}
}
\value{
Formatted PCA plot
}
\description{
Draw a PCA plot with designated PCA components on axis
}
\seealso{
Other PCA functions: 
\code{\link{MDS_plot}()},
\code{\link{PCA_Scree}()},
\code{\link{PCA_biplot}()},
\code{\link{PCA_plot_3d}()},
\code{\link{PCAtools_eigencorplot}()},
\code{\link{get_plot_height}()},
\code{\link{get_plot_width}()},
\code{\link{pc_factor_correlation}()},
\code{\link{t_SNE_plot}()}
}
\concept{PCA functions}
