% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{PCA_plot_3d}
\alias{PCA_plot_3d}
\title{Principal Component Analysis}
\usage{
PCA_plot_3d(
  data,
  sample_info,
  PCAx = 1,
  PCAy = 2,
  PCAz = 3,
  selected_color = "Names",
  selected_shape = "Names"
)
}
\arguments{
\item{data}{Data that has been through pre-processing}

\item{sample_info}{Matrix array with experiment info}

\item{PCAx}{PC on x axis}

\item{PCAy}{PC on y axis}

\item{PCAz}{PC on z axis}

\item{selected_color}{color}

\item{selected_shape}{shape}
}
\value{
Formatted PCA plot
}
\description{
Draw a PCA plot where user selects which PCs on axes
}
