% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{PCA_Scree}
\alias{PCA_Scree}
\title{Principal Component Analysis with PCAtools package}
\usage{
PCA_Scree(processed_data)
}
\arguments{
\item{data}{Matrix of gene data that has been through 
\code{\link{pre_process}()}}
}
\value{
A \code{ggplot} object formatted as a Scree plot
}
\description{
Draw a Scree plot with Horn's and Elbow suggestion for cutoffs using
PCAtools package
}
\seealso{
\code{\link[PCAtools]{screeplot}()} for PCAtools documentation

Other PCA functions: 
\code{\link{MDS_plot}()},
\code{\link{PCA_biplot}()},
\code{\link{PCA_plot_3d}()},
\code{\link{PCA_plot}()},
\code{\link{PCAtools_eigencorplot}()},
\code{\link{pc_factor_correlation}()},
\code{\link{t_SNE_plot}()}
}
\concept{PCA functions}
