% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{pgsea_plot_all}
\alias{pgsea_plot_all}
\title{Pathway analysis with the PGSEA package on all samples}
\usage{
pgsea_plot_all(
  go,
  my_range,
  data,
  select_contrast,
  gene_sets,
  pathway_p_val_cutoff,
  n_pathway_show,
  select_go,
  show_pathway_id,
  margin = c(3, 1, 13, 38),
  plot_colors = NULL
)
}
\arguments{
\item{go}{String designating the section of the database to query for
pathway analysis. See \code{\link{gmt_category}()} for choices.}

\item{my_range}{Vector of the (min_set_size, max_set_size)}

\item{data}{Matrix of gene data that has been through the
\code{\link{pre_process}()}}

\item{select_contrast}{String designating the comparison from DEG analysis to
filter for the significant genes. See the 'comparison' element from the list
returned from \code{\link{limma_value}()} for options.}

\item{gene_sets}{List of vectors with each vector being the
set of genes that correspond to a particular pathway in
the database. See list returned from \code{\link{read_gene_sets}()}}

\item{pathway_p_val_cutoff}{Significant p-value to determine
enriched pathways}

\item{n_pathway_show}{Number of pathways to return in final
result}

\item{select_go}{pathway category.}

\item{show_pathway_id}{Whether to show pathway id for GO and KEGG pathways}

\item{plot_colors}{A vector of colors for activated/surpressed pathways}
}
\value{
A data frame with the results of the pathway analysis.
 The data frame has five columns for the direction of the
 regulation, the pathway description, the stat value, the
 number of overlapping genes, and the p-value.
}
\description{
Run pathway analysis with the PGSEA package using the results
from the \code{\link{limma_value}()} on all samples.
}
\seealso{
Other pathway functions: 
\code{\link{fgsea_data}()},
\code{\link{find_overlap}()},
\code{\link{gage_data}()},
\code{\link{get_gsva_plot_data}()},
\code{\link{get_pathway_list_data}()},
\code{\link{get_pgsea_plot_all_samples_data}()},
\code{\link{get_pgsea_plot_data}()},
\code{\link{gsva_data}()},
\code{\link{kegg_pathway}()},
\code{\link{pathway_select_data}()},
\code{\link{pgsea_data}()},
\code{\link{plot_gsva}()},
\code{\link{plot_pgsea}()},
\code{\link{reactome_data}()},
\code{\link{remove_pathway_id}()},
\code{\link{remove_pathway_id_second}()}
}
\concept{pathway functions}
