% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{plot_data_transform}
\alias{plot_data_transform}
\title{Transform Pathway Data with Plot Output}
\usage{
plot_data_transform(plot_data, contrast, method, genes, org, path_id, go)
}
\arguments{
\item{plot_data}{Pathway data from selected method}

\item{contrast}{Selected contrast from DEG1 tab}

\item{method}{Selected pathway method}

\item{genes}{Gene data}

\item{org}{Selected org from pre-process step}

\item{path_id}{Show pathway ID toggle}

\item{go}{Selected pathway database}
}
\value{
Data frame with pathway names and p-value as columns, along with
         gene expression values from original data
}
\description{
Transform data from plotted Pathway Methods like PGSEA, GSVA, PLAGE, etc.
for data downloading, so p-vals and pathways are columns
}
