% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{plot_deg_scatter}
\alias{plot_deg_scatter}
\title{Scatter plot of the comparison groups}
\usage{
plot_deg_scatter(
  select_contrast,
  comparisons,
  top_genes,
  limma_p_val,
  limma_fc,
  contrast_samples,
  processed_data,
  sample_info
)
}
\arguments{
\item{select_contrast}{String indicating specific comparison from DEG 
analysis to filter for the significant genes}

\item{comparisons}{Vector of the comparisons from the results list
of \code{\link{limma_value}()}}

\item{top_genes}{List of top genes from results list of 
\code{\link{limma_value}()}}

\item{limma_p_val}{Numeric value indicating significant p-value to use to in 
determining the expressed genes}

\item{limma_fc}{Numeric value indicating the minimum fold change value to use 
in determining the expressed genes}

\item{contrast_samples}{Samples that are included in the selected
comparison}

\item{processed_data}{Matrix of gene data that has been through
\code{\link{pre_process}()}}

\item{sample_info}{Matrix of experiment design information for grouping}
}
\value{
A formatted ggplot with the X-axis as the mean expression
 of one contrast group and the Y-axis as the mean expression of
 the other contrast group.
}
\description{
Create a scatter plot of the expression value for each gene
in the two groups for the selected contrast. For the selected
contrast, the mean expression is calculated for a gene in both
group of samples in the contrast and plotted in a scatter plot.
}
\seealso{
Other plots: 
\code{\link{PCA_plot}()},
\code{\link{cor_plot}()},
\code{\link{draw_sample_tree}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{individual_plots}()},
\code{\link{k_means_elbow}()},
\code{\link{mean_sd_plot}()},
\code{\link{plot_ma}()},
\code{\link{plot_volcano}()},
\code{\link{sd_density}()},
\code{\link{t_SNE_plot}()},
\code{\link{total_counts_ggplot}()}

Other DEG functions: 
\code{\link{change_names}()},
\code{\link{deg_heat_data}()},
\code{\link{deg_information}()},
\code{\link{experiment_design_txt}()},
\code{\link{limma_value}()},
\code{\link{list_block_factors_ui}()},
\code{\link{list_factors_ui}()},
\code{\link{list_interaction_terms_ui}()},
\code{\link{list_model_comparisons_ui}()},
\code{\link{mod_label_server}()},
\code{\link{mod_label_ui}()},
\code{\link{plot_ma}()},
\code{\link{plot_volcano}()},
\code{\link{select_reference_levels_ui}()},
\code{\link{volcano_data}()}
}
\concept{DEG functions}
\concept{plots}
