% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{plot_gsva}
\alias{plot_gsva}
\title{Heatmap of GSVA pathway analysis}
\usage{
plot_gsva(
  my_range,
  processed_data,
  contrast_samples,
  gene_sets,
  pathway_p_val_cutoff,
  n_pathway_show,
  select_go,
  show_pathway_id,
  algorithm = "gsva"
)
}
\arguments{
\item{my_range}{Vector of the (min_set_size, max_set_size)}

\item{processed_data}{Matrix of gene data that has been through
\code{\link{pre_process}()}}

\item{contrast_samples}{Sample columns that correspond to the
selected comparison}

\item{gene_sets}{List of vectors with each vector being the
set of genes that correspond to a particular pathway in
the database. See list returned from \code{\link{read_gene_sets}()}}

\item{pathway_p_val_cutoff}{Significant p-value to determine
enriched pathways}

\item{n_pathway_show}{Number of significant pathways to show}

\item{select_go}{pathway category.}

\item{show_pathway_id}{Whether to show pathway id for GO and KEGG pathways}
}
\value{
A heatmap plot with the rows as the significant
 pathways and the columns corresponding to the samples.
}
\description{
Create a heatmap from the pathway analysis using the GSVA
package. The heatmap shows the expression in each group for
each significantly enriched pathway.
}
\seealso{
Other pathway functions: 
\code{\link{fgsea_data}()},
\code{\link{find_overlap}()},
\code{\link{gage_data}()},
\code{\link{get_gsva_plot_data}()},
\code{\link{get_pathway_list_data}()},
\code{\link{get_pgsea_plot_all_samples_data}()},
\code{\link{get_pgsea_plot_data}()},
\code{\link{gsva_data}()},
\code{\link{kegg_pathway}()},
\code{\link{pathway_select_data}()},
\code{\link{pgsea_data}()},
\code{\link{pgsea_plot_all}()},
\code{\link{plot_pgsea}()},
\code{\link{reactome_data}()},
\code{\link{remove_pathway_id_second}()},
\code{\link{remove_pathway_id}()}
}
\concept{pathway functions}
