% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{plot_ma}
\alias{plot_ma}
\title{Plot mean expression and fold change}
\usage{
plot_ma(data, plot_colors, anotate_genes = NULL)
}
\arguments{
\item{data}{Dataframe of processed data from \code{volcano_data()}}

\item{plot_colors}{List containing three colors to differentiate between
the up-regulated, down-regulated, and other genes}

\item{anotate_genes}{List containing the gene names to be labeled on the
plot. Default is NULL.}
}
\value{
A ggplot with the X-axis the mean expression value and
 the Y-axis the calculated fold-change from the DEG analysis.
}
\description{
Draw a ggplot of the overal mean expression for each gene
and the calculated fold change for the selected comparison.
}
