% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{plot_pgsea}
\alias{plot_pgsea}
\title{PLOT PGSEA}
\usage{
plot_pgsea(
  my_range,
  processed_data,
  contrast_samples,
  gene_sets,
  pathway_p_val_cutoff,
  n_pathway_show
)
}
\description{
PLOT PGSEA
}
