% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{plot_upset}
\alias{plot_upset}
\title{Plot upset graph}
\usage{
plot_upset(results)
}
\arguments{
\item{results}{Dataframe from \code{venn_data()}}
}
\value{
A \code{ggplot2} object
}
\description{
Plot upset graph
}
