% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{pre_process}
\alias{pre_process}
\title{Pre-Process the data}
\usage{
pre_process(
  data,
  missing_value = c("geneMedian", "treatAsZero", "geneMedianInGroup"),
  data_file_format = c(1, 2, 3, 4),
  low_filter_fpkm,
  n_min_samples_fpkm,
  log_transform_fpkm,
  log_start_fpkm,
  min_counts,
  n_min_samples_count,
  counts_transform,
  counts_log_start,
  no_fdr
)
}
\arguments{
\item{data}{Matrix of data that has already gone through
\code{\link{convert_data}()}}

\item{missing_value}{String indicating method to deal with missing data. This
should be one of "geneMedian", "treatAsZero", or "geneMedianInGroup"}

\item{data_file_format}{Integer indicating the data format. This should be
one of 1 for read counts data, 2 for normalized expression, 3 for fold
changes with adjusted P-values, or 4 for fold changes without P-values}

\item{low_filter_fpkm}{Integer for low count filter if
\code{data_file_format} is normalized expression, \code{NULL} otherwise}

\item{n_min_samples_fpkm}{Integer for minimum samples if
\code{data_file_format} is normalized expression, \code{NULL} otherwise}

\item{log_transform_fpkm}{TRUE/FALSE if a log transformation should be
applied to normalized expression data}

\item{log_start_fpkm}{Integer added to log transformation if
\code{data_file_format} is normalized expression, \code{NULL} otherwise}

\item{min_counts}{Numeric value for minimum count if
\code{data_file_format} is read counts}

\item{n_min_samples_count}{Integer for minimum libraries with
\code{min_counts} if \code{data_file_format} is read counts}

\item{counts_transform}{Integer to indicate which transformation to make if
\code{data_file_format} is read counts. This should be one of 1 for
log2(CPM+c) (EdgeR), 2 for variance stabilizing transformation (VST), or 3
for regularized log (rlog)}

\item{counts_log_start}{Integer added to log if \code{counts_transform} is
log2(CPM + 2)}

\item{no_fdr}{TRUE/FALSE to indicate fold-changes-only data with no p-values
if \code{data_file_format} is fold changes}
}
\value{
A list containing the transformed data, the mean kurtosis,
the raw counts, a data type warning, the size of the original data,
and p-values.
}
\description{
This function takes in user-defined values to
process the data for the EDA. Processing steps depend on data format, but
generally include missing value imputation, data filtering, and data
transformations.
}
\seealso{
\itemize{
\item \code{\link[edgeR]{cpm}()} for information on calculating counts per
million
\item \code{\link[DESeq2]{vst}()} for information on variance
stabilizing transformation
\item \code{\link[DESeq2]{rlog}()} for
information on the regularized log transformation
}

Other preprocess functions: 
\code{\link{chr_counts_ggplot}()},
\code{\link{chr_normalized_ggplot}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{gene_counts_ggplot}()},
\code{\link{get_x_axis_label_size}()},
\code{\link{individual_plots}()},
\code{\link{mean_sd_plot}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{total_counts_ggplot}()}
}
\concept{preprocess functions}
