% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{prep_cloud_data}
\alias{prep_cloud_data}
\title{Prepare Word Cloud Data}
\usage{
prep_cloud_data(
  gene_lists,
  cluster,
  cluster2 = NULL,
  compare = FALSE,
  cloud_go,
  select_org,
  converted,
  gmt_file,
  idep_data,
  gene_info
)
}
\arguments{
\item{gene_lists}{}

\item{cluster}{}

\item{cluster2}{}

\item{compare}{}

\item{cloud_go}{}

\item{select_org}{}

\item{converted}{}

\item{gmt_file}{}

\item{idep_data}{}

\item{gene_info}{}
}
\value{
Returns data frame of words from pathways in the selected cluster
and their frequencies.
}
\description{
Prepares words in pathway and corresponding frequencies for 
constructing word clouds
}
