% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{prep_venn}
\alias{prep_venn}
\title{Prep data for venn diagram}
\usage{
prep_venn(limma, up_down_regulated, select_comparisons_venn)
}
\arguments{
\item{limma}{Returned list of results from the limma_value
function}

\item{up_down_regulated}{Split the comparisons into either
up or down regulated}

\item{select_comparisons_venn}{The comparisons to plot on the
venn diagram}
}
\value{
A data frame of formatted data for use in a venn diagram or
upset plot
}
\description{
This function prepares the data from the DEG results
with some data transformations and filtering based on specified
parameters. The data is then passed into the \code{plot_venn} and
\code{plot_upset} functions.
}
