% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_09_network.R
\name{prepare_module_csv}
\alias{prepare_module_csv}
\title{prepare_module_csv}
\usage{
prepare_module_csv(wgcna, select_org, all_gene_info)
}
\arguments{
\item{wgcna}{List returned from the \code{get_wgcna}}

\item{select_org}{Organism the expression data is for}

\item{all_gene_info}{Gene info that was found from querying the database}
}
\value{
A dataframe containing for csv file
}
\description{
Create a dataframe from the wgcna object to export as a CSV file, this code is from old idep
}
