% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_09_network.R
\name{prepare_module_csv_filter}
\alias{prepare_module_csv_filter}
\title{prepare_module_csv_filter}
\usage{
prepare_module_csv_filter(module_data, module)
}
\arguments{
\item{module}{The module to create a plot of hte top genes for,}

\item{wgcna}{List returned from the \code{get_wgcna}}
}
\value{
A dataframe containing for csv file
}
\description{
Create a dataframe from the wgcna object to export as a CSV file, allow the user to filter the module
}
