% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{print_dataframe}
\alias{print_dataframe}
\title{Print out a matrix's content as R code to regenerate the matrix
This is used for generating R code from DESeq2 and limma commands
Used mostly for design matrix in limma}
\usage{
print_dataframe(df, df_name = "ma", convert_matrix = FALSE)
}
\arguments{
\item{df}{a matrix or data frame of numbers or characters}

\item{df_name}{a string, specifying the name of the matrix to be print out}

\item{convert_matrix, }{whether to convert to a matrix}
}
\value{
a string. R code that can reconstruct the matrix
}
\description{
Print out a matrix's content as R code to regenerate the matrix
This is used for generating R code from DESeq2 and limma commands
Used mostly for design matrix in limma
}
