% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{print_vector}
\alias{print_vector}
\title{print the content of a vector as R code for the reconstruc of the same vector}
\usage{
print_vector(x, name)
}
\arguments{
\item{x}{vector}

\item{name}{name for the vector
returned}
}
\value{
a string object containing the R code
}
\description{
print the content of a vector as R code for the reconstruc of the same vector
}
