% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_heatmap.R
\name{process_heatmap_data}
\alias{process_heatmap_data}
\title{DATA FOR HEATMAP PANEL}
\usage{
process_heatmap_data(
  data,
  n_genes,
  gene_centering,
  gene_normalize,
  sample_centering,
  sample_normalize
)
}
\description{
DATA FOR HEATMAP PANEL
}
