% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{proper}
\alias{proper}
\title{Capitalize first letter}
\usage{
proper(x)
}
\arguments{
\item{x}{A string}
}
\value{
Cleaned string
}
\description{
This function converts gene set names to a more readable format
}
\examples{
proper("GOBP_mmu_mgi_GO:0000183_chromatin_silencing_at_rDNA")
}
