% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{reactome_data}
\alias{reactome_data}
\title{REACTOME DATA}
\usage{
reactome_data(
  select_contrast,
  my_range,
  limma,
  gene_p_val_cutoff,
  converted,
  idep_data,
  pathway_p_val_cutoff,
  n_pathway_show,
  absolute_fold
)
}
\description{
REACTOME DATA
}
