% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{reactome_gene_list}
\alias{reactome_gene_list}
\title{Find list of genes in Reactome Pathway}
\usage{
reactome_gene_list(sig_pathway, data, gene_info, converted)
}
\arguments{
\item{sig_pathway}{Pathway selection}

\item{data}{ReactomePA pathway dataset}

\item{gene_info}{Gene dataset}

\item{converted}{pre_process converted data}
}
\value{
data frame of genes from the selected pathway
}
\description{
Find list of genes in Reactome Pathway
}
