% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{read_gene_sets}
\alias{read_gene_sets}
\title{Read pathway gene sets}
\usage{
read_gene_sets(converted, all_gene_names, go, select_org, idep_data, my_range)
}
\arguments{
\item{converted}{List of converted gene information  from
\code{\link{convert_id}()}}

\item{all_gene_names}{Data frame of gene names from 
\code{\link{get_all_gene_names}()}}

\item{go}{String designating the section of the database to query for pathway
analysis. See \code{\link{gmt_category}()} for choices.}

\item{select_org}{String designating with organism is being analyzed}

\item{idep_data}{List of data returned from \code{\link{get_idep_data}()}}

\item{my_range}{Vector of the (min_set_size, max_set_size)}
}
\value{
A list with each entry a list of gene IDs that correspond to
 a pathway.
}
\description{
Use the IDs from the converted database to find the
gene sets for all the pathways in the database. Returns
a list with each entry a vector of IDs corresponding to
a description of a pathway in the database.
}
