% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{read_gmt}
\alias{read_gmt}
\title{Read .gmt file}
\usage{
read_gmt(file_path)
}
\arguments{
\item{file_path}{String with path to file}
}
\value{
Data frame with data from gmt file
}
\description{
This functions cleans and converts gene names to upper case in  a .gmt file
}
