% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{read_pathway_sets}
\alias{read_pathway_sets}
\title{Read pathway sets for gene query}
\usage{
read_pathway_sets(
  all_gene_names_query,
  converted,
  go,
  select_org,
  gmt_file,
  idep_data,
  gene_info
)
}
\arguments{
\item{all_gene_names_query}{Subsetted data frame of genes names from
\code{\link{get_all_gene_names}()} to query database with.}

\item{converted}{List of conversion information on the original
IDs returned from the \code{\link{convert_id}()}.}

\item{go}{String designating the section of the database to query for pathway
analysis. See \code{\link{gmt_category}()} for choices.}

\item{select_org}{String designating what organism is being analyzed.}

\item{gmt_file}{For NEW species the gmt file to use
for the pathway analysis}

\item{idep_data}{Data built in to idep}

\item{gene_info}{The gene info from the converted IDs and
the function gene_info()}
}
\value{
This function returns a list with values that are
  used in the find_overlap function. The list contains
  pathway_table which is the overlap and total genes for
  each pathway that is enriched in the query. The list
  also contains the query_set of genes, the total_genes
  number which is used in the calculation of the p-values
  and the pathway files that contain gmt information on
  the mathced species.
}
\description{
This function provides the gene set information
to perform enrichment analysis on the provided
query.
}
