% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{refine_ggplot2}
\alias{refine_ggplot2}
\title{Change ggplot2 plots}
\usage{
refine_ggplot2(p, gridline, ggplot2_theme = "light")
}
\arguments{
\item{p}{ggplot2 object}

\item{gridline}{TRUE of FALSE}
}
\value{
ggplot2 object
}
\description{
Change ggplot2 plots
}
