% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{remove_pathway_id}
\alias{remove_pathway_id}
\title{Remove Pathway ID from pathway name 
Only for GO and KEGG pathways}
\usage{
remove_pathway_id(strings, select_go)
}
\arguments{
\item{strings}{a vector of strings}

\item{select_go}{GOBP, GOCC, GOMP or KEGG or something else}
}
\value{
a vector of strings
}
\description{
Path:hsa00270 Cysteine and methionine metabolism 
          --> Cysteine and methionine metabolism
}
\seealso{
Other pathway functions: 
\code{\link{fgsea_data}()},
\code{\link{find_overlap}()},
\code{\link{gage_data}()},
\code{\link{get_pathway_list_data}()},
\code{\link{get_pgsea_plot_all_samples_data}()},
\code{\link{get_pgsea_plot_data}()},
\code{\link{kegg_pathway}()},
\code{\link{pathway_select_data}()},
\code{\link{pgsea_data}()},
\code{\link{pgsea_plot_all}()},
\code{\link{plot_pgsea}()},
\code{\link{reactome_data}()},
\code{\link{remove_pathway_id_second}()}
}
\concept{pathway functions}
