% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{remove_pathway_id_second}
\alias{remove_pathway_id_second}
\title{Remove Pathway ID from pathway name in PGSEA
Only for GO and KEGG pathways}
\usage{
remove_pathway_id_second(strings, select_go)
}
\arguments{
\item{strings}{a vector of strings}

\item{select_go}{GOBP, GOCC, GOMP or KEGG or something else}
}
\value{
a vector of strings
}
\description{
5.40e-05 Path:hsa04110 Cell cycle
    --> 5.40e-05 Cell cycle
}
\seealso{
Other pathway functions: 
\code{\link{fgsea_data}()},
\code{\link{find_overlap}()},
\code{\link{gage_data}()},
\code{\link{get_gsva_plot_data}()},
\code{\link{get_pathway_list_data}()},
\code{\link{get_pgsea_plot_all_samples_data}()},
\code{\link{get_pgsea_plot_data}()},
\code{\link{gsva_data}()},
\code{\link{kegg_pathway}()},
\code{\link{pathway_select_data}()},
\code{\link{pgsea_data}()},
\code{\link{pgsea_plot_all}()},
\code{\link{plot_gsva}()},
\code{\link{plot_pgsea}()},
\code{\link{reactome_data}()},
\code{\link{remove_pathway_id}()}
}
\concept{pathway functions}
