% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{sd_density}
\alias{sd_density}
\title{Density plot of data standard deviation}
\usage{
sd_density(data, n_genes_max)
}
\arguments{
\item{data}{Data matrix that has been through pre-processing}

\item{n_genes_max}{Integer for the upper limit of gene range}
}
\value{
Formatted density plot of the standard deviation
distribution.
}
\description{
Draw a density plot of the standard deviation in the
data. The function also adds a  vertical red lines for a specified range of 
genes.
}
\seealso{
Other plots: 
\code{\link{draw_sample_tree}()},
\code{\link{heatmap_main}()}

Other clustering: 
\code{\link{draw_sample_tree}()},
\code{\link{heatmap_main}()},
\code{\link{process_heatmap_data}()}
}
\concept{clustering}
\concept{plots}
