% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_07_deg2.R
\name{short_species_names}
\alias{short_species_names}
\title{FUNCTION_TITLE}
\usage{
short_species_names(specie_mame)
}
\arguments{
\item{specie_mame}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
