% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_01_load_data.R
\name{showGeneIDs}
\alias{showGeneIDs}
\title{Show sample gene IDs in modal}
\usage{
showGeneIDs(species, db, nGenes = 10)
}
\arguments{
\item{species}{String indicating selected organism for the expression}

\item{db}{String indicating selected database}

\item{nGenes}{Number of genes to show for each idType}
}
\value{
a data frame with example gene IDs
}
\description{
Show example gene IDs for a selected species.
}
