% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_01_load_data.R
\name{sum_fun}
\alias{sum_fun}
\title{Calculate sum of vector}
\usage{
sum_fun(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a number, or NA if all the numbers are NA
}
\description{
This function refines the sum function to handle missing value, used 
with the aggregate function
}
