% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{transform_comparisons}
\alias{transform_comparisons}
\title{comparisons in all levels of the other factor
"stage: MN vs. EN"  -->  c("MN_AB-EN_AB", "EN_Nodule-EN_AB")}
\usage{
transform_comparisons(comparison, key_model_factors, sample_info)
}
\arguments{
\item{comparison}{Comparison}

\item{key_model_factors}{model factors}

\item{sample_info}{a matrix of experimental design
returned list}
}
\value{
a list of comparisons
}
\description{
comparisons in all levels of the other factor
"stage: MN vs. EN"  -->  c("MN_AB-EN_AB", "EN_Nodule-EN_AB")
}
