% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{transform_comparisons_two_factor_needs_deleted}
\alias{transform_comparisons_two_factor_needs_deleted}
\title{comparisons in all levels of the other factor 
"stage: MN vs. EN"  -->  c("MN_AB-EN_AB", "EN_Nodule-EN_AB")}
\usage{
transform_comparisons_two_factor_needs_deleted(
  comparison,
  key_model_factors,
  sample_info
)
}
\arguments{
\item{comparison}{Comparison}

\item{key_model_factors}{model factors
returned list}
}
\value{
a list of comparisons
}
\description{
comparisons in all levels of the other factor 
"stage: MN vs. EN"  -->  c("MN_AB-EN_AB", "EN_Nodule-EN_AB")
}
