% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{var_imp_plots}
\alias{var_imp_plots}
\title{Principal Component Analysis with PCAtools package
 
 Convert PCA loadings to percent contribution values by squaring and plot
 them}
\usage{
var_imp_plots(data, all_gene_names, PC)
}
\arguments{
\item{all_gene_names}{vector of gene names from 
\code{\link{pre_process}())}}

\item{PC}{# Selected principal component}

\item{processed_data}{Matrix of gene data that has been through
\code{\link{pre_process}())}}
}
\value{
A \code{ggplot} object as a formatted plot generated with PCAtools
 package
}
\description{
Principal Component Analysis with PCAtools package
 
 Convert PCA loadings to percent contribution values by squaring and plot
 them
}
\concept{PCA Functions}
