% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{add_legend}
\alias{add_legend}
\title{Add a legend}
\usage{
add_legend(...)
}
\description{
Add a sample legends to the main heatmap
}
\note{
https://stackoverflow.com/questions/3932038/plot-a-legend-outside-of-the-plotting-area-in-base-graphics
}
