% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{change_names}
\alias{change_names}
\title{FUNCTION_TITLE}
\usage{
change_names(comparison)
}
\arguments{
\item{comparison}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
