% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{change_names}
\alias{change_names}
\title{Change comparison names}
\usage{
change_names(comparison)
}
\arguments{
\item{comparison}{String of comparison to change the name for}
}
\value{
The altered comparison string, see the description for example.
}
\description{
Change comparison names in limma from "KO_ko-WT_ko" to    "KO-WT_for_ko"
}
\seealso{
Other DEG functions: 
\code{\link{deg_heat_data}()},
\code{\link{deg_information}()},
\code{\link{experiment_design_txt}()},
\code{\link{limma_value}()},
\code{\link{list_block_factors_ui}()},
\code{\link{list_factors_ui}()},
\code{\link{list_interaction_terms_ui}()},
\code{\link{list_model_comparisons_ui}()},
\code{\link{mod_label_server}()},
\code{\link{mod_label_ui}()},
\code{\link{plot_deg_scatter}()},
\code{\link{plot_ma}()},
\code{\link{plot_volcano}()},
\code{\link{select_reference_levels_ui}()},
\code{\link{volcano_data}()}
}
\concept{DEG functions}
