% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{clean_gene_set}
\alias{clean_gene_set}
\title{Data cleaning on gene sets}
\usage{
clean_gene_set(gene_set)
}
\arguments{
\item{gene_set}{Vector of genes to be cleans.}
}
\value{
Vector of cleaned genes.
}
\description{
Clean up a set of genes by removing duplicates, spaces, and other control characters from
gene names
}
\note{
This function is called internally in other idepGolem functions.
}
