% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{clean_query}
\alias{clean_query}
\title{Clean gene query}
\usage{
clean_query(query_input)
}
\arguments{
\item{query_input}{Vector with list of genes}
}
\value{
Vector of cleaned gene list
}
\description{
Clean and prepare a gene query by adding tabs and new lines after each gene
}
\note{
This function is called internally in other idepGolem functions.
}
