% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{convert_ensembl}
\alias{convert_ensembl}
\title{FUNCTION_TITLE}
\usage{
convert_ensembl(query, species, idep_date, convert_type = "entrez")
}
\arguments{
\item{query}{DESCRIPTION.}

\item{species}{DESCRIPTION.}

\item{idep_date}{DESCRIPTION.}

\item{convert_type}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
