% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{convert_id}
\alias{convert_id}
\title{convert_id This function takes gene IDs and converts them to ensembl data.}
\usage{
convert_id(query, idep_data, select_org = "BestMatch")
}
\arguments{
\item{query}{A character vector of gene IDs}

\item{idep_data}{A instance of the output from get_idep_data
(link to documentation)}

\item{select_org}{A character of the species that wants to be looked up,
default to \code{"BestMatch"}}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
