% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{deg_information}
\alias{deg_information}
\title{Information on results from DEG analysis 
returns results from DEG analysis}
\usage{
deg_information(
  limma_value,
  gene_names,
  processed_data,
  no_id_conversion = FALSE
)
}
\arguments{
\item{limma_value}{List of results from \code{\link{limma_value}()}}

\item{gene_names}{List of gene IDs}

\item{processed_data}{Processed gene data from \code{\link{pre_process}()}}

\item{no_id_conversion}{TRUE/FALSE. If true, gene ids will not be mapped to Ensembl}
}
\value{
Dataframe of DEG analysis results
}
\description{
Information on results from DEG analysis 
returns results from DEG analysis
}
\seealso{
Other DEG functions: 
\code{\link{change_names}()},
\code{\link{deg_heat_data}()},
\code{\link{experiment_design_txt}()},
\code{\link{limma_value}()},
\code{\link{list_block_factors_ui}()},
\code{\link{list_factors_ui}()},
\code{\link{list_interaction_terms_ui}()},
\code{\link{list_model_comparisons_ui}()},
\code{\link{mod_label_server}()},
\code{\link{mod_label_ui}()},
\code{\link{plot_deg_scatter}()},
\code{\link{plot_ma}()},
\code{\link{plot_volcano}()},
\code{\link{select_reference_levels_ui}()},
\code{\link{volcano_data}()}
}
\concept{DEG functions}
