% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{detect_groups}
\alias{detect_groups}
\title{Detect groups by sample names}
\usage{
detect_groups(sample_names, sample_info = NULL)
}
\arguments{
\item{sample_names}{Vector of column headings from data file or design file}

\item{sample_info}{Matrix of the experiment design information}
}
\value{
A character vector with the groups
}
\description{
Detects groups from column names in sample info file so that they can be used
for things such as coloring plots or building the model for DEG analysis.
}
\note{
This function is mainly called internally in other idepGolem functions.
}
