% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{eda_boxplot}
\alias{eda_boxplot}
\title{Boxplot for processed data}
\usage{
eda_boxplot(processed_data, sample_info)
}
\arguments{
\item{processed_data}{Matrix of data that has gone through
\code{\link{pre_process}()}}

\item{sample_info}{Matrix of experiment design information}
}
\value{
Boxplot of the distribution of counts for each sample as a
 \code{ggplot} object.
}
\description{
This function takes the processed data and creates
a boxplot of number of sequences mapped to each
tissue sample.
}
\seealso{
Other plots: 
\code{\link{PCA_plot}()},
\code{\link{chr_counts_ggplot}()},
\code{\link{cor_plot}()},
\code{\link{draw_sample_tree}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{individual_plots}()},
\code{\link{k_means_elbow}()},
\code{\link{mean_sd_plot}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{sd_density}()},
\code{\link{t_SNE_plot}()},
\code{\link{total_counts_ggplot}()}

Other preprocess functions: 
\code{\link{chr_counts_ggplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{individual_plots}()},
\code{\link{mean_sd_plot}()},
\code{\link{pre_process}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{total_counts_ggplot}()}
}
\concept{plots}
\concept{preprocess functions}
