% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{experiment_design_txt}
\alias{experiment_design_txt}
\title{Create a string of the model design}
\usage{
experiment_design_txt(
  sample_info,
  select_factors_model,
  select_block_factors_model,
  select_interactions
)
}
\arguments{
\item{sample_info}{Matrix of experiment design information for grouping}

\item{select_factors_model}{Vector of strings indicating the selected factors
for the model expression, see \code{\link{list_factors_ui}()} for options}

\item{select_block_factors_model}{Vector of strings indicating selected 
factors for batch effect, see \code{\link{list_block_factors_ui}()} for 
options}

\item{select_interactions}{Vector of strings indicating selected the 
interaction terms being used in the model design, see 
\code{\link{list_interaction_terms_ui}()} for options}
}
\value{
Returns a string of the model design being used for
 the DEG analysis
}
\description{
Use the model design selections to create a string of the
model design being used for the DEG analysis.
}
\seealso{
Other DEG functions: 
\code{\link{change_names}()},
\code{\link{deg_heat_data}()},
\code{\link{deg_information}()},
\code{\link{limma_value}()},
\code{\link{list_block_factors_ui}()},
\code{\link{list_factors_ui}()},
\code{\link{list_interaction_terms_ui}()},
\code{\link{list_model_comparisons_ui}()},
\code{\link{mod_label_server}()},
\code{\link{mod_label_ui}()},
\code{\link{plot_deg_scatter}()},
\code{\link{plot_ma}()},
\code{\link{plot_volcano}()},
\code{\link{select_reference_levels_ui}()},
\code{\link{volcano_data}()}
}
\concept{DEG functions}
