% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{extract_word}
\alias{extract_word}
\title{Extract a word}
\usage{
extract_word(word_list)
}
\arguments{
\item{word_list}{word list}
}
\value{
extracted words
}
\description{
Extract a word
}
\examples{
# ADD_EXAMPLES_HERE
}
