% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{fgsea_data}
\alias{fgsea_data}
\title{Pathway analysis with the FGSEA package}
\usage{
fgsea_data(
  select_contrast,
  my_range,
  limma,
  gene_p_val_cutoff,
  gene_sets,
  absolute_fold,
  pathway_p_val_cutoff,
  n_pathway_show
)
}
\arguments{
\item{select_contrast}{String designating the comparison from DEG analysis to
filter for the significant genes. See the 'comparison' element from the list
returned from \code{\link{limma_value}()} for options.}

\item{my_range}{Vector of the (min_set_size, max_set_size)}

\item{limma}{Results list from the \code{\link{limma_value}()}}

\item{gene_p_val_cutoff}{Significant p-value to filter
the top genes fold change by}

\item{gene_sets}{List of vectors with each vector being the
set of genes that correspond to a particular pathway in
the database. See results list from
\code{\link{read_gene_sets}()}.}

\item{absolute_fold}{TRUE/FALSE to use the absolute value of the fold
change}

\item{pathway_p_val_cutoff}{Significant p-value to determine
enriched pathways}

\item{n_pathway_show}{Number of pathways to return in final
result}
}
\value{
A data frame with the results of the pathway analysis.
 The data frame has five columns for the direction of the
 regulation, the pathway description, the stat value, the
 number of overlapping genes, and the p-value.
}
\description{
Run pathway analysis with the FGSEA package using the results
from the \code{\link{limma_value}()}.
}
\seealso{
Other pathway functions: 
\code{\link{find_overlap}()},
\code{\link{gage_data}()},
\code{\link{get_pathway_list_data}()},
\code{\link{get_pgsea_plot_all_samples_data}()},
\code{\link{get_pgsea_plot_data}()},
\code{\link{kegg_pathway}()},
\code{\link{pathway_select_data}()},
\code{\link{pgsea_data}()},
\code{\link{pgsea_plot_all}()},
\code{\link{plot_pgsea}()},
\code{\link{reactome_data}()},
\code{\link{remove_pathway_id_second}()},
\code{\link{remove_pathway_id}()}
}
\concept{pathway functions}
