% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_random.R
\name{find_overlap}
\alias{find_overlap}
\title{FIND OVERLAP FUNCTION
Main function. Find a query set of genes enriched with functional category}
\usage{
find_overlap(
  converted,
  all_gene_names,
  g_info,
  GO,
  selectOrg,
  minFDR,
  reduced = FALSE
)
}
\description{
FIND OVERLAP FUNCTION
Main function. Find a query set of genes enriched with functional category
}
