% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{gene_counts_ggplot}
\alias{gene_counts_ggplot}
\title{Creates a barplot of the count of genes by type}
\usage{
gene_counts_ggplot(
  counts_data,
  sample_info,
  type = "",
  all_gene_info,
  plots_color_select
)
}
\arguments{
\item{counts_data}{Matrix of expression values used to align genes with
metadata (values themselves are ignored)}

\item{sample_info}{Matrix of experiment design information for grouping
samples}

\item{type}{String designating the type of data to be used in the title.
Commonly either "Raw" or "Transformed"}

\item{all_gene_info}{Gene info, including chr., gene type etc.}

\item{plots_color_select}{Vector of colors for plots}
}
\value{
A barplot as a \code{ggplot} object
}
\description{
This function works with either raw count or normalized expression matrices.
It links the supplied gene metadata to tally how many genes fall into each
gene type and returns the counts as a bar chart.
}
\seealso{
Other preprocess functions: 
\code{\link{chr_counts_ggplot}()},
\code{\link{chr_normalized_ggplot}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{get_x_axis_label_size}()},
\code{\link{individual_plots}()},
\code{\link{mean_sd_plot}()},
\code{\link{pre_process}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{total_counts_ggplot}()}

Other plots: 
\code{\link{PCA_plot}()},
\code{\link{chr_counts_ggplot}()},
\code{\link{chr_normalized_ggplot}()},
\code{\link{cor_plot}()},
\code{\link{draw_sample_tree}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{individual_plots}()},
\code{\link{k_means_elbow}()},
\code{\link{mean_sd_plot}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{sd_density}()},
\code{\link{t_SNE_plot}()},
\code{\link{total_counts_ggplot}()}
}
\concept{plots}
\concept{preprocess functions}
