% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_random.R
\name{get_gene_info}
\alias{get_gene_info}
\title{Retrieve detailed gene information}
\usage{
get_gene_info(converted, select_org, gene_info_files)
}
\arguments{
\item{converted}{List of converted gene information  from
\code{\link{convert_id}()}}

\item{select_org}{String designating the organism being analyzed.}

\item{gene_info_files}{List of gene files from the 'gene_info_files' element
of the result list from \code{\link{get_idep_data}()}.}
}
\value{
A data frame
}
\description{
Retrieve detailed gene information
}
