% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{get_plot_height}
\alias{get_plot_height}
\title{Gets plot height size from session$clientdata}
\usage{
get_plot_height(client_data, plot_name, tab)
}
\arguments{
\item{client_data}{session info from shiny}

\item{plot_name}{Name of plot object}

\item{tab}{Tab name}
}
\value{
height in inches of plot
}
\description{
Gets plot height size from session$clientdata
}
