% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{get_plot_width}
\alias{get_plot_width}
\title{Gets plot width dimensions from session$clientdata}
\usage{
get_plot_width(client_data, plot_name, tab)
}
\arguments{
\item{client_data}{session info from shiny}

\item{plot_name}{Name of plot object}

\item{tab}{Tab name}
}
\value{
string with dimension info
}
\description{
Gets plot width dimensions from session$clientdata
}
