% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{gmt_category}
\alias{gmt_category}
\title{Database choices for the converted IDs}
\usage{
gmt_category(converted, converted_data, select_org, gmt_file, idep_data)
}
\arguments{
\item{converted}{List of converted gene information  from
\code{\link{convert_id}()}}

\item{converted_data}{Data matrix element with the converted ensembl_IDs from
the list returned from \code{\link{convert_data}()}}

\item{select_org}{String designating the organism being analyzed}

\item{gmt_file}{For NEW species the gmt file to use
for the pathway analysis, otherwise \code{NULL}}

\item{idep_data}{List of data returned from \code{\link{get_idep_data}()}}
}
\value{
A character vector of names of the section of data to perform
  pathway analysis on
}
\description{
This function provides a list of the portions of the
database to query that contain genes from the converted
IDs.
}
