% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_heatmap.R
\name{heatmap_main}
\alias{heatmap_main}
\title{HEATMAP FUNCTION}
\usage{
heatmap_main(
  data,
  n_genes,
  heatmap_cutoff,
  sample_info,
  select_factors_heatmap,
  dist_funs,
  dist_function,
  hclust_function,
  no_sample_clustering,
  heatmap_color_select
)
}
\description{
HEATMAP FUNCTION
}
