% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{heatmap_main}
\alias{heatmap_main}
\title{Draw a heatmap of processed data}
\usage{
heatmap_main(
  data,
  cluster_meth,
  heatmap_cutoff,
  sample_info,
  select_factors_heatmap,
  dist_funs,
  dist_function,
  hclust_function,
  sample_clustering,
  heatmap_color_select,
  row_dend,
  k_clusters,
  re_run,
  selected_genes
)
}
\arguments{
\item{data}{Data matrix returned from \code{\link{process_heatmap_data}()}}

\item{cluster_meth}{Integer indicating which clustering method to use 1 for 
hierarchical and 2 for kmeans.}

\item{heatmap_cutoff}{Numeric value for Z score max to filter data}

\item{sample_info}{Matrix of experiment design information from load data}

\item{select_factors_heatmap}{Factor group for annotation legend}

\item{dist_funs}{List of distance functions to use in heatmap}

\item{dist_function}{The selected distance function to use}

\item{hclust_function}{Type of clustering to perform}

\item{sample_clustering}{TRUE/FALSE Specify whether to cluster columns}

\item{heatmap_color_select}{Vector of colors for heatmap scale}

\item{row_dend}{TRUE/FALSE Hide row dendrogram,}

\item{k_clusters}{Number of clusters to use for k-means}

\item{re_run}{Integer for a seed to Re-run k-means with}

\item{selected_genes}{Character list of genes to label on the heatmap}
}
\value{
Heatmap of the processed data.
}
\description{
Uses the package ComplexHeatmaps to draw a heatmap of the
processed data that has been prepared for the heatmap. The
returned heatmap visualizes the processed expression of the
gene range provided in \code{\link{process_heatmap_data}()}.
}
\seealso{
\code{\link{dist_functions}()} for available distance functions, 
 \code{\link{hcluster_functions}()} for available functions for hierarchical 
 clustering

Other clustering: 
\code{\link{cluster_heat_click_info}()},
\code{\link{draw_sample_tree}()},
\code{\link{heat_sub}()},
\code{\link{k_means_elbow}()},
\code{\link{process_heatmap_data}()},
\code{\link{sd_density}()},
\code{\link{sub_heat_ann}()}

Other heatmaps: 
\code{\link{cluster_heat_click_info}()},
\code{\link{heat_sub}()},
\code{\link{process_heatmap_data}()},
\code{\link{sub_heat_ann}()}
}
\concept{clustering}
\concept{heatmaps}
