% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{individual_plots}
\alias{individual_plots}
\title{Individual plotting function for genes}
\usage{
individual_plots(
  individual_data,
  sample_info,
  selected_gene,
  gene_plot_box,
  use_sd,
  lab_rotate
)
}
\arguments{
\item{individual_data}{Data that has been merged with the gene info}

\item{sample_info}{Matrix of experiment design information for
grouping samples}

\item{gene_plot_box}{TRUE/FALSE for individual sample plot or grouped
data plot}

\item{use_sd}{TRUE/FALSE for standard error or standard deviation bars on
bar plot}

\item{lab_rotate}{Numeric value indicating what angle to rotate
the x-axis labels}

\item{select_gene}{List of gene(s) to be plotted}
}
\value{
A \code{ggplot} object. For \code{gene_plot_box = TRUE} the return
 will be a lineplot for the expression of each individual sample for the
 selected gene. If \code{gene_plot_box = FALSE} the return will be a barplot
 for the groups provided in the sample information.
}
\description{
Takes in the merged data and other data to provide
plots on individual gene names. Depening on the selections
this function will either return a barplot from the
grouped data or a lineplot from the individual sample.
}
\seealso{
\code{\link{merge_data}()}

Other plots: 
\code{\link{PCA_plot}()},
\code{\link{cor_plot}()},
\code{\link{draw_sample_tree}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{k_means_elbow}()},
\code{\link{mean_sd_plot}()},
\code{\link{plot_deg_scatter}()},
\code{\link{plot_ma}()},
\code{\link{plot_volcano}()},
\code{\link{sd_density}()},
\code{\link{t_SNE_plot}()},
\code{\link{total_counts_ggplot}()}

Other preprocess functions: 
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{mean_sd_plot}()},
\code{\link{pre_process}()},
\code{\link{total_counts_ggplot}()}
}
\concept{plots}
\concept{preprocess functions}
