% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{kegg_pathway_id}
\alias{kegg_pathway_id}
\title{Find pathway IDs for a KEGG description}
\usage{
kegg_pathway_id(
  pathway_description,
  Species,
  GO = "KEGG",
  select_org,
  gmt_files,
  org_info,
  idep_data
)
}
\arguments{
\item{pathway_description}{String providing description of a pathway in the
iDEP database, see the results from \code{\link{read_gene_Sets}()} for 
options}

\item{Species}{String designating the organism being analyzed}

\item{GO}{String designating the portion of the database to query, 
for this function, it should be "KEGG"}

\item{select_org}{String designating the organism being analyzed}

\item{gmt_files}{GMT files from the iDEP database}

\item{org_info}{Organism information files from the iDEP
database}

\item{idep_data}{List of data from \code{\link{get_idep_data}()}}
}
\value{
Return the KEGG ID for the pathway.
}
\description{
From a pathway description find the KEGG ID.
}
\seealso{
This function is used internally in \code{\link{kegg_pathway}()}
}
