% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_deg1.R
\name{list_block_factors_ui}
\alias{list_block_factors_ui}
\title{LIST BLOCK FACTORS UI}
\usage{
list_block_factors_ui(
  sample_info,
  select_factors_model,
  data_file_format,
  deg_method,
  id
)
}
\description{
LIST BLOCK FACTORS UI
}
