% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_deg1.R
\name{list_interaction_terms_ui}
\alias{list_interaction_terms_ui}
\title{LIST INTERACTION TERMS}
\usage{
list_interaction_terms_ui(sample_info, select_factors_model, id)
}
\description{
LIST INTERACTION TERMS
}
