% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_deg1.R
\name{list_model_comparisons_ui}
\alias{list_model_comparisons_ui}
\title{MODEL COMPARISONS UI}
\usage{
list_model_comparisons_ui(
  sample_info,
  select_factors_model,
  processed_data,
  id
)
}
\description{
MODEL COMPARISONS UI
}
