% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{loose_rock_proper}
\alias{loose_rock_proper}
\title{Capitalizes all words in string}
\usage{
loose_rock_proper(x)
}
\arguments{
\item{x}{String}
}
\value{
a capitalized string (all words)

#examples
loose_rock_proper("i saw a dEaD parrot")
}
\description{
from: https://github.com/averissimo/loose.rock/blob/master/R/string.R
}
