% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{MDS_plot}
\alias{MDS_plot}
\title{MDS plot}
\usage{
MDS_plot(data, sample_info, selected_shape, selected_color)
}
\arguments{
\item{data}{Matrix of gene data that has been through pre-processing}

\item{sample_info}{Matrix of sample information from experiment design file}

\item{selected_shape}{String designating factor to add shape to points by}

\item{selected_color}{String designating factor to color points by}
}
\value{
Formatted PCA plot
}
\description{
Draw a multidimensional scaling (MDS) plot
}
\seealso{
Other PCA functions: 
\code{\link{PCA_Scree}()},
\code{\link{PCA_biplot}()},
\code{\link{PCA_plot_3d}()},
\code{\link{PCA_plot}()},
\code{\link{PCAtools_eigencorplot}()},
\code{\link{get_plot_height}()},
\code{\link{get_plot_width}()},
\code{\link{pc_factor_correlation}()},
\code{\link{t_SNE_plot}()}
}
\concept{PCA functions}
