% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{mod_label_server}
\alias{mod_label_server}
\title{Server component to customize gene labels}
\usage{
mod_label_server(id, data_list, method = c("volcano", "ma"))
}
\arguments{
\item{id}{String designating namespace ID}

\item{data_list}{List of gene data from \code{volcano_data()}}

\item{method}{String designating if the results are for the volcano plot or
ma plot. Should be one of "volcano" or "ma"}
}
\value{
A shiny server function.
}
\description{
This component contains the pop-up modal and data processing to customize
 how genes are labeled on the volcano or ma plot. Users have the option to
 not label genes, label specific genes, label top n genes by a certain value,
 or label genes above certain threshold.
}
\seealso{
\code{\link{mod_label_ui}()}

Other shiny modules: 
\code{\link{mod_label_ui}()}

Other DEG functions: 
\code{\link{change_names}()},
\code{\link{deg_heat_data}()},
\code{\link{deg_information}()},
\code{\link{experiment_design_txt}()},
\code{\link{limma_value}()},
\code{\link{list_block_factors_ui}()},
\code{\link{list_factors_ui}()},
\code{\link{list_interaction_terms_ui}()},
\code{\link{list_model_comparisons_ui}()},
\code{\link{mod_label_ui}()},
\code{\link{plot_deg_scatter}()},
\code{\link{plot_ma}()},
\code{\link{plot_volcano}()},
\code{\link{select_reference_levels_ui}()},
\code{\link{volcano_data}()}
}
\concept{DEG functions}
\concept{shiny modules}
