% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{mod_label_ui}
\alias{mod_label_ui}
\title{UI component to customize gene labels}
\usage{
mod_label_ui(id)
}
\arguments{
\item{id}{String designating the namespace ID}
}
\value{
A shiny UI function
}
\description{
This component contains an action button to activate the pop-up modal to
customize how genes are labeled on the volcano or ma plot.
}
\seealso{
\code{\link{mod_label_server}()}

Other shiny modules: 
\code{\link{mod_label_server}()}

Other DEG functions: 
\code{\link{change_names}()},
\code{\link{deg_heat_data}()},
\code{\link{deg_information}()},
\code{\link{experiment_design_txt}()},
\code{\link{limma_value}()},
\code{\link{list_block_factors_ui}()},
\code{\link{list_factors_ui}()},
\code{\link{list_interaction_terms_ui}()},
\code{\link{list_model_comparisons_ui}()},
\code{\link{mod_label_server}()},
\code{\link{plot_deg_scatter}()},
\code{\link{plot_ma}()},
\code{\link{plot_volcano}()},
\code{\link{select_reference_levels_ui}()},
\code{\link{volcano_data}()}
}
\concept{DEG functions}
\concept{shiny modules}
