% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{pathway_data_transform}
\alias{pathway_data_transform}
\title{Transform Pathway Data}
\usage{
pathway_data_transform(data, contrast, method, genes, org, path_id, go)
}
\arguments{
\item{data}{Pathway data from selected method}

\item{contrast}{Selected contrast from DEG1 tab}

\item{method}{Selected pathway method}

\item{genes}{Gene data}

\item{org}{Selected org from pre-process step}

\item{path_id}{Show pathway ID toggle}

\item{go}{Selected pathway database}
}
\value{
Data frame with hyperlinks and urls for pathways
}
\description{
Transform data from various pathway methods for uniform download format
}
