% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{pc_factor_correlation}
\alias{pc_factor_correlation}
\title{Correlations Between Principle Components and Factors}
\usage{
pc_factor_correlation(data, sample_info)
}
\arguments{
\item{data}{Matrix of gene data that has been through
\code{\link{pre_process}()}}

\item{sample_info}{Matrix of sample information from experiment design file}
}
\value{
A string
}
\description{
This function calculates the correlation between the principle components
 and factors and return the result in a formatted string.
}
\seealso{
Other PCA functions: 
\code{\link{MDS_plot}()},
\code{\link{PCA_Scree}()},
\code{\link{PCA_biplot}()},
\code{\link{PCA_plot_3d}()},
\code{\link{PCA_plot}()},
\code{\link{PCAtools_eigencorplot}()},
\code{\link{t_SNE_plot}()}
}
\concept{PCA functions}
