% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{PCA_plot_3d}
\alias{PCA_plot_3d}
\title{Principal component analysis 3D plot}
\usage{
PCA_plot_3d(
  data,
  sample_info,
  PCAx = 1,
  PCAy = 2,
  PCAz = 3,
  selected_color = "Names",
  selected_shape = "Names"
)
}
\arguments{
\item{data}{Matrix of gene data that has been through
\code{\link{pre_process}()}}

\item{sample_info}{Matrix of sample information from experiment design file}

\item{PCAx}{Integer designating the PC to be plotted on the x axis}

\item{PCAy}{Integer designating the PC to be plotted on the y axis}

\item{PCAz}{Integer designating the PC to be plotted on the z axis}

\item{selected_color}{String designating factor to color points by. Should be
one of the design factors from the design file or "Names" as default which
automatically detects groups from gene data file}

\item{selected_shape}{String designating factor to shape points by.
Should be one of the design factors from the design file or "Names" as
default which automatically detects groups from gene data file}
}
\value{
Formatted PCA plot
}
\description{
Draw a 3D PCA plot with designated PCA components on axis
}
\seealso{
\code{\link{PCA_plot}()}

Other PCA functions: 
\code{\link{MDS_plot}()},
\code{\link{PCA_Scree}()},
\code{\link{PCA_biplot}()},
\code{\link{PCA_plot}()},
\code{\link{PCAtools_eigencorplot}()},
\code{\link{pc_factor_correlation}()},
\code{\link{t_SNE_plot}()}
}
\concept{PCA functions}
