% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{PCA_Scree}
\alias{PCA_Scree}
\title{Principal Component Analysis with PCAtools package}
\usage{
PCA_Scree(processed_data)
}
\arguments{
\item{data}{Data that has been through pre-processing}
}
\value{
Formatted Scree plot using PCAtools package
}
\description{
Draw a Scree plot with Horn's and Elbow suggestion for cutoffs using PCAtools package
}
\seealso{
Other PCA functions: 
\code{\link{MDS_plot}()},
\code{\link{PCA_biplot}()},
\code{\link{PCA_plot_3d}()},
\code{\link{PCA_plot}()},
\code{\link{PCAtools_eigencorplot}()},
\code{\link{get_plot_height}()},
\code{\link{get_plot_width}()},
\code{\link{pc_factor_correlation}()},
\code{\link{t_SNE_plot}()}
}
\concept{PCA functions}
