% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{PCAtools_eigencorplot}
\alias{PCAtools_eigencorplot}
\title{Principal Component Analysis with PCAtools package}
\usage{
PCAtools_eigencorplot(processed_data, sample_info)
}
\arguments{
\item{sample_info}{Matrix of sample information from experiment design file}

\item{data}{Matrix of gene data that has been through
\code{\link{pre_process}())}}
}
\value{
A \code{ggplot} object as a formatted plot generated with PCAtools
 package
}
\description{
Generates a plot showing correlations between Principal Components and design factors
}
\seealso{
\code{\link[PCAtools]{pca}()} and
  \code{\link[PCAtools]{eigencorplot}()} for specific documentation from the
  PCAtools package

Other PCA functions: 
\code{\link{MDS_plot}()},
\code{\link{PCA_Scree}()},
\code{\link{PCA_biplot}()},
\code{\link{PCA_plot_3d}()},
\code{\link{PCA_plot}()},
\code{\link{pc_factor_correlation}()},
\code{\link{t_SNE_plot}()}
}
\concept{PCA functions}
