% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{plot_ma}
\alias{plot_ma}
\title{Plot mean expression and fold change}
\usage{
plot_ma(data, plot_colors, anotate_genes = NULL)
}
\arguments{
\item{data}{Dataframe of processed data from \code{\link{volcano_data}()}}

\item{plot_colors}{List containing three colors to differentiate between
the up-regulated, down-regulated, and other genes}

\item{anotate_genes}{List containing the gene names to be labeled on the
plot. Default is NULL.}
}
\value{
A ggplot with the X-axis the mean expression value and
 the Y-axis the calculated fold-change from the DEG analysis.
}
\description{
Draw a ggplot of the overal mean expression for each gene
and the calculated fold change for the selected comparison.
}
\seealso{
\code{\link{volcano_data}()} for data prep

Other DEG functions: 
\code{\link{change_names}()},
\code{\link{deg_heat_data}()},
\code{\link{deg_information}()},
\code{\link{experiment_design_txt}()},
\code{\link{limma_value}()},
\code{\link{list_block_factors_ui}()},
\code{\link{list_factors_ui}()},
\code{\link{list_interaction_terms_ui}()},
\code{\link{list_model_comparisons_ui}()},
\code{\link{mod_label_server}()},
\code{\link{mod_label_ui}()},
\code{\link{plot_deg_scatter}()},
\code{\link{plot_volcano}()},
\code{\link{select_reference_levels_ui}()},
\code{\link{volcano_data}()}

Other plots: 
\code{\link{PCA_plot}()},
\code{\link{cor_plot}()},
\code{\link{draw_sample_tree}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{individual_plots}()},
\code{\link{k_means_elbow}()},
\code{\link{mean_sd_plot}()},
\code{\link{plot_deg_scatter}()},
\code{\link{plot_volcano}()},
\code{\link{sd_density}()},
\code{\link{t_SNE_plot}()},
\code{\link{total_counts_ggplot}()}
}
\concept{DEG functions}
\concept{plots}
