% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano DEG plot}
\usage{
plot_volcano(data, anotate_genes = NULL, plot_colors)
}
\arguments{
\item{data}{Dataframe of processed data from \code{volcano_data()}.}

\item{anotate_genes}{List of gene names to anotate. Default is NULL.}

\item{plot_colors}{List containing three colors to differentiate between
the up-regulated, down-regulated, and other genes}
}
\value{
ggplot with the fold value as the X-axis and the log 10
 value of the adjusted p-value as the Y-axis.
}
\description{
Use the results from limma-value to create a volcano plot to
illustrate the significantly expressed genes. Up and down
regulated genes are colored on the ggplot.
}
