% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{prep_download}
\alias{prep_download}
\title{Prep heatmap data for download}
\usage{
prep_download(heatmap, heatmap_data, cluster_meth)
}
\arguments{
\item{cluster_meth}{}
}
\description{
Prep heatmap data for download by merging with clusters.
}
