% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{read_gene_sets}
\alias{read_gene_sets}
\title{READ GENE SETS}
\usage{
read_gene_sets(
  all_gene_names,
  converted,
  go,
  select_org,
  gmt_range,
  gmt_file,
  idep_data
)
}
\description{
READ GENE SETS
}
