% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{read_gmt}
\alias{read_gmt}
\title{read_gmt}
\usage{
read_gmt(file_path)
}
\arguments{
\item{file_path}{string with path to file}
}
\value{
data frame with data from gmt file
}
\description{
This functions cleans and converts to upper case genes in .gmt file to upper case
}
\examples{
# ADD_EXAMPLES_HERE
}
