% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{read_pathway_sets}
\alias{read_pathway_sets}
\title{READ GENE SETS}
\usage{
read_pathway_sets(
  all_gene_names_query,
  converted,
  go,
  select_org,
  gmt_range,
  gmt_file,
  idep_data,
  gene_info
)
}
\description{
READ GENE SETS
}
