% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{refine_ggplot2}
\alias{refine_ggplot2}
\title{Change ggplot2 plots}
\usage{
refine_ggplot2(p, gridline, ggplot2_theme = "light")
}
\arguments{
\item{p}{\code{ggplot} object}

\item{gridline}{TRUE/FALSE to indicate adding gridlines}

\item{ggplot2_theme}{String to indicate which \code{ggplot} theme to use. 
Should be one of "linedraw", "classic", "gray", "light", "dark", or "bw"}
}
\value{
Formatted \code{ggplot} object
}
\description{
Change ggplot2 plots
}
