% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{sd_density}
\alias{sd_density}
\title{Density plot of data standard deviation}
\usage{
sd_density(data, n_genes_max)
}
\arguments{
\item{data}{Data matrix that has been through pre-processing}

\item{n_genes_max}{Integer for the upper limit of gene range}
}
\value{
Formatted density plot of the standard deviation
distribution.
}
\description{
Draw a density plot of the standard deviation in the
data. The function also adds a  vertical red lines for a specified range of
genes.
}
\seealso{
Other plots: 
\code{\link{PCA_plot}()},
\code{\link{chr_counts_ggplot}()},
\code{\link{chr_normalized_ggplot}()},
\code{\link{cor_plot}()},
\code{\link{draw_sample_tree}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{gene_counts_ggplot}()},
\code{\link{individual_plots}()},
\code{\link{k_means_elbow}()},
\code{\link{mean_sd_plot}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{t_SNE_plot}()},
\code{\link{total_counts_ggplot}()}

Other clustering functions: 
\code{\link{cluster_heat_click_info}()},
\code{\link{create_marker_legend}()},
\code{\link{draw_sample_tree}()},
\code{\link{heat_sub}()},
\code{\link{k_means_elbow}()},
\code{\link{process_heatmap_data}()},
\code{\link{sub_heat_ann}()}
}
\concept{clustering functions}
\concept{plots}
