% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_deg1.R
\name{select_reference_levels_ui}
\alias{select_reference_levels_ui}
\title{SELECT REFERENCE LEVELS}
\usage{
select_reference_levels_ui(
  sample_info,
  select_factors_model,
  data_file_format,
  counts_deg_method,
  id
)
}
\description{
SELECT REFERENCE LEVELS
}
