% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{t_SNE_plot}
\alias{t_SNE_plot}
\title{t-SNE plot}
\usage{
t_SNE_plot(
  data,
  sample_info,
  selected_color = "Names",
  selected_shape = "Names"
)
}
\arguments{
\item{data}{Matrix of gene data that has been through
\code{\link{pre_process}()}}

\item{sample_info}{Matrix of sample information from experiment design file}

\item{selected_color}{String designating factor to color points by. Should be
one of the design factors from the design file or "Names" as default which
automatically detects groups from gene data file}

\item{selected_shape}{String designating factor to shape points by.
Should be one of the design factors from the design file or "Names" as
default which automatically detects groups from gene data file}
}
\value{
A \code{ggplot} object formatted t-SNE plot
}
\description{
Draw a t-distributed stochastic neighbor embedding (t-SNE) plot
}
\seealso{
Other PCA functions: 
\code{\link{MDS_plot}()},
\code{\link{PCA_Scree}()},
\code{\link{PCA_biplot}()},
\code{\link{PCA_plot_3d}()},
\code{\link{PCA_plot}()},
\code{\link{PCAtools_eigencorplot}()},
\code{\link{pc_factor_correlation}()}

Other plots: 
\code{\link{PCA_plot}()},
\code{\link{chr_counts_ggplot}()},
\code{\link{chr_normalized_ggplot}()},
\code{\link{cor_plot}()},
\code{\link{draw_sample_tree}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{individual_plots}()},
\code{\link{k_means_elbow}()},
\code{\link{mean_sd_plot}()},
\code{\link{rRNA_counts_ggplot}()},
\code{\link{sd_density}()},
\code{\link{total_counts_ggplot}()}
}
\concept{PCA functions}
\concept{plots}
