% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{total_counts_ggplot}
\alias{total_counts_ggplot}
\title{Creates a barplot of the count data}
\usage{
total_counts_ggplot(counts_data, sample_info, type = "")
}
\arguments{
\item{counts_data}{Matrix of raw counts from gene expression data}

\item{sample_info}{Matrix of experiment design information for grouping
samples}

\item{type}{String designating the type of data to be used in the title.
Commonly either "Raw" or "Transformed"}
}
\value{
A barplot as a \code{ggplot} object
}
\description{
This function takes in either raw count or processed data and creates a
formatted barplot as a \code{ggplot} object that shows the number
of genes mapped to each sample in millions. This function is only used for
read counts data.
}
\seealso{
Other preprocess functions: 
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{individual_plots}()},
\code{\link{mean_sd_plot}()},
\code{\link{pre_process}()}

Other plots: 
\code{\link{PCA_plot}()},
\code{\link{cor_plot}()},
\code{\link{draw_sample_tree}()},
\code{\link{eda_boxplot}()},
\code{\link{eda_density}()},
\code{\link{eda_scatter}()},
\code{\link{individual_plots}()},
\code{\link{k_means_elbow}()},
\code{\link{mean_sd_plot}()},
\code{\link{plot_deg_scatter}()},
\code{\link{plot_ma}()},
\code{\link{plot_volcano}()},
\code{\link{sd_density}()},
\code{\link{t_SNE_plot}()}
}
\concept{plots}
\concept{preprocess functions}
