% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{volcano_data}
\alias{volcano_data}
\title{Data processing for volcano plot}
\usage{
volcano_data(select_contrast, comparisons, top_genes, limma_p_val, limma_fc)
}
\arguments{
\item{select_contrast}{Comparison from DEG analysis to filter
for the significant genes}

\item{comparisons}{The comparisons vector from the results list
of the limma_value function}

\item{top_genes}{top_genes list from results list of the
limma_value function}

\item{limma_p_val}{Significant p-value to use to in determining
the expressed genes}

\item{limma_fc}{Minimum fold change value to use in determining
the expressed genes}
}
\value{
A list containing processed data for volcano plot in
 \code{plot_volcano()} and list of differently expressed genes for labeling
}
\description{
Data processing for volcano plot
}
